refactor(i18n): #1032 i18n.ts を領域別サブ辞書 (shell/canvas/settings) へ分割
- 1942 行のフラット辞書を lib/i18n/ 配下の ja/en x shell/canvas/settings の 6 サブ辞書へ分割 (全ファイル 500 行以下)
- useT / translate 等の logic は lib/i18n/index.ts へ移動 (import パスは '../lib/i18n' のまま不変)
- 分割の不変条件 (サブ辞書間キー重複なし / ja・en 同一キー集合) を i18n-dicts.test.ts で固定
- 分割前後で 758 キー x 2 言語の key/value 完全一致をスクリプト検証済み
