--- src/renderer/src/lib/i18n.ts
@@ -1,1942 +1,0 @@
-import type { Language } from '../../../types/shared';
-import { useSettings } from './settings-context';
-
-/**
- * フラットキー方式の軽量 i18n。
- * `{param}` 形式のパラメータ置換を最低限サポート。
- */
-type Dict = Record<string, string>;
-
-const ja: Dict = {
-  // ---------- Common ----------
-  'common.close': '閉じる',
-  'common.cancel': 'キャンセル',
-
-  // ---------- Toolbar ----------
-  'toolbar.restart.title': 'アプリを再起動',
-  'toolbar.palette.title': 'コマンドパレット (Ctrl+Shift+P)',
-  'toolbar.settings.title': '設定 (Ctrl+,)',
-
-  // ---------- Window controls (Issue #260 PR-2: カスタムタイトルバー) ----------
-  'windowControls.minimize': '最小化',
-  'windowControls.maximize': '最大化',
-  'windowControls.restore': '元のサイズに戻す',
-  'windowControls.close': '閉じる',
-  'windowControls.group': 'ウィンドウ操作',
-
-  // ---------- Topbar (redesign shell) ----------
-  'topbar.mode.canvas': 'Canvas',
-
-  // ---------- Status bar ----------
-  'status.branch': 'ブランチ',
-  'status.changes': '変更',
-  'status.lang': '言語',
-  'status.theme': 'テーマ',
-  'status.mascot.idle': '待機中',
-  'status.mascot.sleep': 'おやすみ中…',
-  'status.mascot.working': 'エージェント実行中',
-  'status.mascot.thinking': '応答待ち',
-  'status.mascot.done': '完了!',
-  'status.mascot.error': '対応が必要',
-  'status.mascot.excited': 'やる気!',
-
-  // ---------- Mascot section (SettingsModal の「キャラクター」セクション) ----------
-  // Issue #729: MascotSection の isJa 三項 / settings-options.ts hardcode を i18n.ts に集約
-  'settings.mascot.title': 'キャラクター',
-  'settings.mascot.pickTitle': '相棒にする画像を選択',
-  'settings.mascot.imageFilterName': '画像',
-  'settings.mascot.choose': '画像を選ぶ…',
-  'settings.mascot.clear': 'クリア',
-  'settings.mascot.hint':
-    'PNG / GIF (アニメ可) / APNG / WebP / SVG を選べます。\n小さめ (32〜128px) の正方形が綺麗に出ます。',
-  'mascot.desc.vibe': '既定の小さな相棒',
-  'mascot.desc.spark': '明るめで軽い印象',
-  'mascot.desc.mono': '端末になじむ角ばった見た目',
-  'mascot.desc.coder': 'PCでカタカタ作業する相棒',
-  'mascot.desc.custom': '自分で用意した画像 (PNG/GIF/SVG/WebP) を相棒として使う',
-
-  // ---------- Canvas HUD ----------
-  'canvas.apiAgent.teamRole': 'チームロール',
-  'canvas.apiAgent.teamRolePlaceholder': '例: reviewer',
-  'canvas.apiChat.placeholder': '指示を入力してください (/でコマンド、@でエージェントを選択)',
-  'canvas.apiChat.typing': '{name} が入力中…',
-  'canvas.apiChat.ready': '準備完了。指示を入力してください。',
-  'canvas.apiChat.loadingPrompt': 'システムプロンプトを読み込み中…',
-  'canvas.apiChat.configure': '設定でこの API エージェントを構成してください。',
-  'canvas.apiChat.mention': 'エージェントを選択 (@)',
-  'canvas.apiChat.attach': 'ファイル添付 (近日対応)',
-  'canvas.apiChat.send': '送信',
-  'canvas.apiChat.stop': '停止',
-  'canvas.apiChat.cmd.planDesc': '計画を立てる',
-  'canvas.apiChat.cmd.statusDesc': '状態を確認',
-  'canvas.apiChat.cmd.contextDesc': 'コンテキストを表示',
-  'canvas.apiChat.cmd.clearDesc': '履歴をクリア',
-  'canvas.hud.stage': 'ステージ',
-  'canvas.hud.list': 'リスト',
-  'canvas.hud.focus': 'フォーカス',
-  'canvas.hud.fit': 'フィット',
-  'canvas.hud.zoomIn': 'ズームイン',
-  'canvas.hud.zoomOut': 'ズームアウト',
-  'canvas.hud.arrange.open': '整理',
-  // Issue #368: ホバー時の機能役割説明 (Label — 役割)
-  'canvas.hud.stage.tooltip': 'ステージ — エージェントを放射状に並べたビューに切替',
-  'canvas.hud.list.tooltip': 'リスト — エージェントを縦並びの一覧で表示',
-  'canvas.hud.focus.tooltip': 'フォーカス — 選択中のエージェントだけを大きく表示',
-  'canvas.hud.fit.tooltip': 'フィット — Canvas 上の全カードが収まるよう自動で拡縮',
-  'canvas.hud.zoomIn.tooltip': 'ズームイン — Canvas を拡大表示',
-  'canvas.hud.zoomOut.tooltip': 'ズームアウト — Canvas を縮小表示',
-  'canvas.hud.arrange.open.tooltip': '整理 — カードの整頓・サイズ統一・間隔をまとめて調整',
-  'canvas.hud.arrange.tidy': '整頓',
-  'canvas.hud.arrange.unifySize': 'サイズ統一',
-  'canvas.hud.arrange.gap.label': '間隔',
-  'canvas.hud.arrange.gap.tight': '狭い',
-  'canvas.hud.arrange.gap.normal': '標準',
-  'canvas.hud.arrange.gap.wide': '広い',
-
-  // ---------- AppMenu ----------
-  'appMenu.title': 'プロジェクトメニュー',
-  'appMenu.new': '新規プロジェクト…',
-  'appMenu.newHint': '空フォルダを作成/選択',
-  'appMenu.openFolder': 'フォルダを開く…',
-  'appMenu.openFolderHint': '既存のプロジェクト',
-  'appMenu.openFile': 'ファイルを開く…',
-  'appMenu.openFileHint': '単独ファイル',
-  'appMenu.newDialogTitle': '新規プロジェクト',
-  'appMenu.openFolderDialogTitle': 'フォルダを開く',
-  'appMenu.openFileDialogTitle': 'ファイルを開く',
-  'project.newDialogTitle': '新規プロジェクト: 空フォルダを選択/作成',
-  'project.openExistingDialogTitle': '既存プロジェクトを開く',
-  'project.loading': 'プロジェクト読み込み中…',
-  'project.loadError': '読み込みエラー: {error}',
-  'project.initError': '初期化エラー: {error}',
-  'project.newFolderNotEmpty': 'フォルダが空ではありません。既存として開きます',
-  'project.created': '新規プロジェクトを作成',
-  'project.fileParentLoaded': '{file} の親フォルダをプロジェクトとして読み込みました',
-  'project.recentCleared': '最近のプロジェクト履歴をクリアしました',
-  'appMenu.addWorkspaceDialogTitle': 'ワークスペースに追加',
-  'appMenu.addToWorkspace': 'フォルダをワークスペースに追加…',
-  'appMenu.addToWorkspaceHint': 'サイドバーに別ルートを並べる',
-  'appMenu.recent': '最近のプロジェクト',
-  'appMenu.recentCount': '{count} 件の履歴',
-  'appMenu.workspace': 'ワークスペース',
-  'appMenu.clear': 'クリア',
-  'appMenu.empty': '履歴なし',
-  'menubar.file': 'ファイル',
-  'menubar.view': '表示',
-  'menubar.help': 'ヘルプ',
-  'menubar.toggleSidebar': 'サイドバーを切替',
-  'menubar.toggleCanvas': 'IDE / Canvas を切替',
-  'menubar.openPalette': 'コマンドパレット',
-  'menubar.openSettings': '設定…',
-  'menubar.openGithub': 'GitHub で開く',
-  'menubar.restart': '再起動',
-  // ---------- UserMenu (サイドバー左下) ----------
-  'userMenu.settings': '設定',
-  'userMenu.language': '言語',
-  'userMenu.theme': 'テーマ',
-  'userMenu.releases': 'GitHub でリリースを見る',
-  // ---------- ワークスペース (Issue #4) ----------
-  'workspace.roots': 'ワークスペース',
-  'workspace.add': 'フォルダを追加',
-  'workspace.remove': 'ワークスペースから外す',
-  'workspace.removePrimaryConfirm': '{name} を現在のプロジェクトから外します。よろしいですか？',
-  'workspace.removed': '{name} をワークスペースから外しました',
-  'workspace.added': '{name} をワークスペースに追加しました',
-  'workspace.alreadyAdded': '{name} は既に追加されています',
-
-  // ---------- Sidebar ----------
-  'sidebar.files': 'ファイル',
-  'sidebar.changes': '変更',
-  'sidebar.history': '履歴',
-  'sidebar.loading': '読み込み中…',
-  'sidebar.notGitRepo': 'Git リポジトリではありません',
-  'sidebar.noChanges': '変更なし',
-  'sidebar.noSessions': 'このプロジェクトのセッション履歴はまだありません',
-  'sidebar.filesChanged': '{count} 変更',
-  'sidebar.sessionCount': '{count} セッション',
-  'sidebar.refresh': '更新',
-  'sidebar.teams': 'チーム',
-  'sidebar.singleSessions': '個別セッション',
-  'sidebar.notes': 'メモ',
-  'rail.primaryNav': 'メインナビゲーション',
-
-  // ---------- Notes (Issue #17) ----------
-  'notes.title': 'メモ',
-  'notes.placeholder': 'ターミナル間で受け渡したい内容を書き留めてください…\n自動保存されます。',
-  'notes.copy': 'クリップボードにコピー',
-  'notes.clear': 'メモをクリア',
-  'notes.copied': 'メモをコピーしました',
-  'notes.copyFailed': 'コピーに失敗しました',
-  'notes.confirmClear': 'メモをクリアしますか？',
-  'notes.autoSaved': '自動保存済み',
-  'notes.chars': '文字',
-
-  // ---------- Team history ----------
-  'teamHistory.resume': 'チーム「{name}」を復元',
-  'teamHistory.resumed': 'チーム「{name}」を復元しました',
-  'teamHistory.alreadyOpen': 'チーム「{name}」は既に Canvas 上にあります',
-  'teamHistory.delete': '履歴から削除',
-
-  // ---------- File tree / Editor ----------
-  'filetree.refresh': '再読込',
-  'filetree.treeLabel': 'ファイルツリー',
-  'diff.loading': 'diff を読み込み中…',
-  'diff.selectFile': '差分を表示するファイルを選択してください',
-  'diff.error': 'エラー: {error}',
-  'diff.binary': 'バイナリファイルは diff 表示できません: {path}',
-  'diff.new': '(新規追加)',
-  'diff.deleted': '(削除)',
-  'diff.toggleMode': '差分表示モード切替',
-  'diff.toggleInline': 'インラインに切替',
-  'diff.toggleSideBySide': 'サイドバイサイドに切替',
-  'editor.loading': 'ファイルを読み込み中…',
-  'editor.save': '保存 (Ctrl+S)',
-  'editor.save.ariaLabel': '保存',
-  'editor.viewPreview': 'プレビュー表示',
-  'editor.viewSource': 'ソース表示',
-  'editor.binaryNotice': 'バイナリファイルは編集できません: {path}',
-  'editor.nonUtf8Warning':
-    '非 UTF-8 として読み込みました ({path}) — 保存すると元のエンコーディングを失うため編集不可にしています。',
-  'editor.nonUtf8SaveBlocked': '保存は無効化されています (非 UTF-8): {path}',
-  'editor.nonUtf8ReadOnly': '読み取り専用 (非 UTF-8)',
-  'editor.externalChangeConfirm':
-    '{path} は開いた後にディスク上で更新されています。このまま保存すると外部の変更を上書きします。続行しますか?',
-  'editor.saveAborted': '保存を中止しました: {path}',
-  'editor.saved': '保存しました: {path}',
-  'editor.saveFailed': '保存失敗: {error}',
-  'editor.discardSingle': '未保存の変更があります。このファイルを閉じますか？\n\n{path}',
-  'editor.discardMultiple': '未保存の変更があります。このまま切り替えると {count} 個のファイルの変更が失われます。続行しますか？',
-  'editor.restartConfirm': '未保存の変更があります。このままアプリを再起動すると変更が失われます。続行しますか？',
-  // Issue #595: Canvas 上の EditorCard を × / Clear で閉じる際に未保存編集を確認するダイアログ。
-  'editor.confirmDiscardChanges':
-    '未保存の編集が残っています。このカードを閉じると編集内容は失われます。続行しますか？\n\n{path}',
-  'editor.confirmDiscardChangesPlural':
-    '未保存の編集が {count} 件残っています。これらのカードを閉じると編集内容はすべて失われます。続行しますか？\n\n{paths}',
-
-  // ---------- Welcome ----------
-  'welcome.subtitle': 'vibe coding with Claude Code',
-  'welcome.hint1Key': '右',
-  'welcome.hint1Text': 'のターミナルで Claude Code に話しかける',
-  'welcome.hint2Key': '変更',
-  'welcome.hint2Text': 'タブから Claude が触ったファイルの diff を確認',
-  'welcome.hint3Key': '履歴',
-  'welcome.hint3Text': 'タブから過去のセッションに復帰',
-  'welcome.hint4Text': 'でコマンドパレット',
-
-  // ---------- Context menu ----------
-  'ctxMenu.openDiff': '差分を開く',
-  'ctxMenu.reviewDiff': '差分レビューを Claude Code に依頼',
-  'ctxMenu.copyPath': 'パスをコピー',
-  // Issue #251: ファイルツリー右クリックメニュー
-  'ctxMenu.copyAbsolutePath': '絶対パスをコピー',
-  'ctxMenu.copyRelativePath': '相対パスをコピー',
-  'ctxMenu.copyFileName': 'ファイル名をコピー',
-  'ctxMenu.revealInFolder': 'エクスプローラーで開く',
-  // Issue #592: VS Code 互換のファイル/フォルダ操作
-  'ctxMenu.newFile': '新しいファイル',
-  'ctxMenu.newFolder': '新しいフォルダ',
-  'ctxMenu.rename': '名前の変更',
-  'ctxMenu.delete': '削除',
-  'ctxMenu.cut': '切り取り',
-  'ctxMenu.copy': 'コピー',
-  'ctxMenu.paste': '貼り付け',
-  'ctxMenu.duplicate': '複製を作成',
-  'filetree.prompt.newFileName': '新しいファイル名',
-  'filetree.prompt.newFolderName': '新しいフォルダ名',
-  'filetree.prompt.renameTo': '新しい名前',
-  'filetree.confirmDeleteFile': '"{name}" をゴミ箱に移動しますか？',
-  'filetree.confirmDeleteFolder': '"{name}" とその中身をすべてゴミ箱に移動しますか？',
-  'filetree.confirmDeletePermanent': '"{name}" を完全に削除しますか？この操作は元に戻せません。',
-  'filetree.preloadRestartRequired': 'アプリを再起動してください（preload 更新のため）',
-  'canvasMenu.lockTeam': 'チームで一緒に動かす',
-  'canvasMenu.unlockTeam': 'チーム固定を解除',
-  'canvasMenu.deleteCard': 'カードを削除',
-  'canvasMenu.addClaudeHere': 'ここに Claude を追加',
-  'canvasMenu.addCodexHere': 'ここに Codex を追加',
-  'canvasMenu.addCustomAgentHere': 'ここに {name} を追加',
-  'canvasMenu.addFileTreeHere': 'ここにファイルツリーを追加',
-  'canvasMenu.addChangesHere': 'ここに Git 変更を追加',
-  'canvasMenu.addEditorHere': 'ここに空のエディタを追加',
-  'canvasMenu.spawnDefaultTeam': '既定チームを起動',
-
-  // ---------- Claude Code panel ----------
-  'claudePanel.title': 'IDEモード',
-  'claudePanel.notFound.title': 'Claude Code が見つかりません',
-  'claudePanel.notFound.body':
-    '`claude` コマンドが PATH 上に見つかりませんでした。Claude Code をインストールするか、設定で起動コマンドのパスを指定してください。',
-  'claudePanel.notFound.step1Title': 'CLI をインストール',
-  'claudePanel.notFound.step1Desc': '`claude` コマンドがターミナルから実行できる状態にします。',
-  'claudePanel.notFound.step2Title': '設定を確認',
-  'claudePanel.notFound.step2Desc': 'カスタムコマンドを使う場合は Settings から起動コマンドを見直します。',
-  'claudePanel.notFound.installLink': 'Claude Code をインストール',
-  'claudePanel.notFound.retry': '再検出',
-  'claudePanel.notFound.settings': '設定で指定',
-  'claudePanel.checking': '確認中…',
-  'claudePanel.newTab': '新しいターミナルタブ',
-  'claudePanel.addClaude': 'Claude Code を追加',
-  'claudePanel.addCodex': 'Codex を追加',
-
-  // ---------- Team ----------
-  'team.closeTeamConfirm': 'これはチームリーダーです。チーム全体を閉じますか？',
-  'team.closeTeam': 'チームを閉じる',
-  'team.closeLeaderOnly': 'リーダーのみ閉じる',
-
-  // ---------- Canvas ----------
-  'canvas.spawnTeam': 'チーム起動',
-  'canvas.spawnTeam.tooltip': 'チーム起動 — 既定プリセットでリーダー＋メンバーを一括起動',
-  'canvas.spawnTeamMore': 'その他のチーム…',
-  'canvas.spawnTeamMore.tooltip': 'プリセット選択 — 別の組み込みプリセットや最近使ったチームから選ぶ',
-  'canvas.preset': 'プリセット',
-  'canvas.preset.leaderClaude': 'Leader のみで起動 (Claude Code)',
-  'canvas.preset.leaderCodex': 'Leader のみで起動 (Codex)',
-  'canvas.preset.builtinHeader': '組み込み',
-  'canvas.preset.savedHeader': '保存済み',
-  'canvas.preset.leaderCustom': 'Leader のみで起動 ({name})',
-  'canvas.recent': '最近使ったチーム',
-  'canvas.noRecentTeams': '最近使ったチームはありません。プリセットから起動してください。',
-  'canvas.clear': 'クリア',
-  'canvas.clear.tooltip': 'クリア — Canvas 上のカードをすべて削除',
-  'canvas.clearConfirm': 'Canvas 上のカードをすべて削除しますか？',
-  // Issue #595: Clear 実行時に dirty な EditorCard が居ればファイル名一覧と件数を表示する。
-  'canvas.clearConfirmWithDirtyEditors':
-    'Canvas 上のカードをすべて削除します。未保存の編集が {count} 件あり、これらは破棄されます。続行しますか？\n\n{paths}',
-  'canvas.switchToIde': 'IDE モードに戻る',
-  'canvas.switchToIde.tooltip': 'IDE — エディタとターミナル中心の IDE モードへ切替',
-  'canvas.modeToggle': 'Canvas モードに切り替え',
-  'canvas.card.editor': 'エディタ',
-  'canvas.list.title': 'チーム',
-  'canvas.list.empty': 'まだエージェントが配置されていません',
-
-  // ---------- Agent Card ----------
-  'agentCard.close': 'カードを閉じる',
-  'agentCard.confirmCloseTeam':
-    'このカードを閉じると、同じチーム「{name}」のメンバー {count} 名すべて (Leader 含む) が一緒に閉じられます。続行しますか？',
-  'handoff.create': '引き継ぎ',
-  'handoff.createTooltip':
-    '引き継ぎ書を保存し、Leader 自身に MCP で新 Leader 採用 → 交代を依頼します',
-  'handoff.created': '引き継ぎ書 {file} を保存し、Leader に MCP 手順を伝えました',
-  'handoff.action.reveal': '保存先を開く',
-  'handoff.error.noProject':
-    'プロジェクトルートが未設定です。サイドバーからフォルダを開いてからもう一度押してください。',
-  'handoff.error.createFailed': '引き継ぎ書の作成に失敗しました: {detail}',
-  'handoff.error.notLeader': '引き継ぎは Leader カードからのみ開始できます',
-  'handoff.error.injectFailed': 'Leader の PTY への手順注入に失敗しました: {detail}',
-  // Issue #511: PTY inject 失敗の警告 + 手動リトライ
-  'injectFailure.title': '配信失敗 ({code}): {message}',
-  'injectFailure.retry': '再送信',
-  'injectFailure.retryBusy': '再送信中…',
-  'injectFailure.retrySuccess': 'メッセージを再送信しました',
-  'injectFailure.retryFailed': '再送信に失敗しました ({reason})',
-  'injectFailure.retryError': '再送信中にエラーが発生しました: {detail}',
-  'injectFailure.dismiss': '閉じる',
-  // Issue #509: 配送済みだが team_read で確認していない message の表示
-  'inboxUnread.label': '未読 {count} 件 ({ageSec}s 経過)',
-  'inboxUnread.tooltip':
-    'この agent は配送済みのメッセージ {count} 件を {ageSec} 秒間 team_read で確認していません。60s 超過時は督促を検討してください。',
-  'agentStatus.idle': '待機中',
-  'agentStatus.thinking': '思考中',
-  'agentStatus.typing': '応答中',
-
-  // Issue #521: Agent カード 3 行サマリ
-  'agentCard.summary.region': 'エージェントの状態サマリ',
-  'agentCard.summary.noTask': '現在のタスクは未割当',
-  'agentCard.summary.needsLeader': 'Leader の入力待ち',
-  'agentCard.summary.ago.unobserved': '出力はまだ観測されていません',
-  'agentCard.summary.ago.now': '直前に出力',
-  'agentCard.summary.ago.sec': '最終出力から {value} 秒前',
-  'agentCard.summary.ago.min': '最終出力から {value} 分前',
-  'agentCard.summary.ago.hour': '最終出力から {value} 時間前',
-  'agentCard.summary.ago.day': '最終出力から {value} 日前',
-
-  // Issue #510: Agent カード health badge (TeamHub diagnostics 由来)
-  'agentCard.summary.health.state.alive': '稼働中',
-  'agentCard.summary.health.state.stale': '沈黙中',
-  'agentCard.summary.health.state.dead': '応答なし',
-  'agentCard.summary.health.state.unknown': '不明',
-  'agentCard.summary.health.silent.sec': '{state} ({value} 秒沈黙)',
-  'agentCard.summary.health.silent.min': '{state} ({value} 分沈黙)',
-  'agentCard.summary.health.tooltip': 'Health: {state} / 直近自己申告: {status}',
-  'agentCard.summary.health.noStatus': '自己申告なし',
-
-  // Issue #521: Canvas 全体サマリ HUD
-  'canvas.hud.summary.label': 'Canvas 全体の状態サマリ',
-  'canvas.hud.summary.active': '進行中',
-  'canvas.hud.summary.active.tooltip': '進行中 — 直近に出力があったエージェントの数',
-  'canvas.hud.summary.blocked': 'Leader 待ち',
-  'canvas.hud.summary.blocked.tooltip':
-    'Leader 待ち — Leader の入力 / handoff ack を待っているエージェントの数',
-  'canvas.hud.summary.stale': '停滞',
-  'canvas.hud.summary.stale.tooltip': '停滞 — 5 分以上出力が無いエージェントの数',
-  'canvas.hud.summary.completed': '完了',
-  'canvas.hud.summary.completed.tooltip': '完了 — handoff ack 済 / 退役済のエージェントの数',
-  'canvas.hud.summary.dead': '応答なし',
-  'canvas.hud.summary.dead.tooltip':
-    '応答なし — 15 分以上 PTY 出力なしのエージェントの数 (Hub diagnostics 由来)',
-
-  // Issue #522: Team Presets panel
-  'preset.title': 'チームプリセット',
-  'preset.button.tooltip': 'プリセット — 現在のチーム編成を保存・再構築',
-  'preset.saveCurrent': '現在のチームを保存',
-  'preset.saveCurrent.tooltip': '今 Canvas に並んでいる Agent カードをプリセットとして保存',
-  'preset.save': '保存',
-  'preset.name': '名前',
-  'preset.namePlaceholder': '例: 計画 + 実装 + レビュー チーム',
-  'preset.description': '説明',
-  'preset.descriptionPlaceholder': '任意のメモ (どんな課題に向く編成か等)',
-  'preset.apply': '適用',
-  'preset.apply.tooltip': 'このプリセットの役職構成を Canvas に展開',
-  'preset.delete': '削除',
-  'preset.delete.tooltip': 'このプリセットをディスクから削除',
-  'preset.empty': '保存されたプリセットはまだありません',
-  'preset.loading': '読み込み中…',
-  'preset.roleCount': '{count} 名',
-  'preset.saved': 'プリセット「{name}」を保存しました',
-  'preset.applied': '「{name}」のメンバー {count} 名を Canvas に追加しました',
-  'preset.deleted': 'プリセット「{name}」を削除しました',
-  'preset.error.empty': 'Canvas に Agent カードがありません。先にチームを組んでから保存してください',
-  'preset.error.noName': 'プリセット名を入力してください',
-  'preset.error.listFailed': 'プリセット一覧の読み込みに失敗しました',
-  'preset.error.saveFailed': 'プリセット保存に失敗しました: {detail}',
-  'preset.error.deleteFailed': 'プリセット削除に失敗しました: {detail}',
-
-  // Issue #514: Team Dashboard
-  'dashboard.title': 'チームダッシュボード',
-  'dashboard.button.tooltip': 'チームダッシュボード — 全メンバーの状態 / タスク / 経過を一覧',
-  'dashboard.count': '{count} 名',
-  'dashboard.col.member': 'メンバー',
-  'dashboard.col.state': '状態',
-  'dashboard.col.task': '担当タスク',
-  'dashboard.col.lastSeen': '最終出力',
-  'dashboard.state.active': '進行中',
-  'dashboard.state.blocked': 'Leader 待ち',
-  'dashboard.state.stale': '停滞',
-  'dashboard.state.completed': '完了',
-  'dashboard.state.idle': '待機',
-  'dashboard.task.unassigned': 'タスク未割り当て',
-  'dashboard.lastSeen.never': '未観測',
-  'dashboard.empty.noTeam':
-    '対象のチームが Canvas にありません。Agent カードを 1 枚以上配置してください',
-  'dashboard.empty.noMembers':
-    'このチームにはまだメンバーがいません。Leader から `team_recruit` でメンバーを招集してください',
-  'dashboard.banner.humanGate': 'Human gate が blocked: Leader の判断待ちです',
-  'dashboard.alert.leaderInput': 'Leader 入力待ち',
-  'dashboard.alert.staleOutput': '5 分以上出力なし',
-  // Issue #615: dual / multi preset 対応の team section heading
-  'dashboard.team.label': 'チーム {index}',
-
-  // ---------- Sessions ----------
-  'sessions.resume': 'セッション {id} に戻る',
-  'sessions.messages': '{count} 件',
-  // Issue #837: messageCount が走査上限で打ち切られたときの "N+" 表示。
-  'sessions.messagesCapped': '{count}+ 件',
-  'sessions.loadMore': '残り {remaining} 件を表示',
-
-  // ---------- Tab ----------
-  'tab.pinned': 'ピン留め中',
-  'tab.newOutput': '新しい出力',
-  'tab.pin': 'ピン留め',
-  'tab.unpin': 'ピンを外す',
-  'tab.close': 'タブを閉じる',
-  'tab.closeWithShortcut': '閉じる (Ctrl+W)',
-  'fonts.family': 'フォントファミリ',
-  'fonts.custom': '（カスタム）',
-  'fonts.size': 'サイズ (px)',
-  'fonts.customCss': 'カスタム CSS font-family',
-
-  // ---------- Roles ----------
-
-  // ---------- Settings ----------
-  'settings.title': '設定',
-  // Issue #729: WelcomePane の inline isJa を i18n.ts に移管
-  'welcome.title': '静かな集中で、すばやく進める。',
-  'welcome.recentProjects': '最近のプロジェクト',
-  'welcome.recentProjectsTitle': 'すぐに戻れる作業面',
-  'welcome.workspaceLabel': 'ワークスペース',
-  'welcome.quickStart': 'クイックスタート',
-  'welcome.quickStartTitle': 'よく使う操作',
-  // Issue #729: canvas-layout-helpers の language ベース hardcode を i18n.ts に移管
-  'canvas.agentCount': '{count} エージェント',
-  'canvas.orgAgentCount': '{organizationCount} 組織 / {agentCount} エージェント',
-  // Issue #729: RoleProfilesSection の isJa 三項を i18n.ts に移管
-  'settings.roles.title': 'ロール定義',
-  'settings.roles.desc':
-    'vibe-team のメンバーロールを定義します。Leader が team_recruit で動的に呼ぶときの選択肢になります。',
-  'settings.roles.globalPreamble': '全エージェント共通の前置き',
-  'settings.roles.globalPreambleHint': '全 system prompt の先頭に挿入',
-  'settings.roles.confirmDelete': '"{id}" を削除しますか?',
-  'settings.roles.addCustom': 'カスタムロールを追加',
-  'settings.roles.newCustomDesc': '新しいカスタムロール。',
-  'settings.roles.builtin': '組み込み',
-  'settings.roles.custom': 'カスタム',
-  'settings.roles.color': '色',
-  'settings.roles.glyph': 'グリフ',
-  'settings.roles.defaultEngine': '既定エンジン',
-  'settings.roles.permissions': '権限',
-  'settings.roles.promptEn': 'システムプロンプト (EN)',
-  'settings.roles.promptJa': 'システムプロンプト (JA)',
-  'settings.roles.promptPlaceholders':
-    'placeholder: {teamName} {selfLabel} {selfDescription} {roster} {tools} {globalPreamble}',
-  'settings.roles.deleteRole': 'このロールを削除',
-  // Issue #729: settings-section-meta.tsx の FIXED_LABELS_JA を i18n.ts へ移管
-  'settings.section.general.label': '一般',
-  'settings.section.general.title': '一般',
-  'settings.section.general.desc': '言語と密度設定',
-  'settings.section.appearance.label': '表示',
-  'settings.section.appearance.title': '表示',
-  'settings.section.appearance.desc': 'テーマ、配色、キャラクター',
-  'settings.section.fonts.label': 'フォント',
-  'settings.section.fonts.title': 'フォント',
-  'settings.section.fonts.desc': 'UI / エディタ / ターミナルのフォント',
-  'settings.section.claude.label': 'Claude Code',
-  'settings.section.claude.title': 'Claude Code',
-  'settings.section.claude.desc': '起動コマンドと引数',
-  'settings.section.codex.label': 'Codex',
-  'settings.section.codex.title': 'Codex',
-  'settings.section.codex.desc': '起動コマンドと引数',
-  // Issue #1068: codex team_send の配送方式トグル
-  'settings.codexDelivery.title': 'team_send の配送方式',
-  'settings.codexDelivery.label': '配送方式',
-  'settings.codexDelivery.optBackend': 'バックエンド (app-server) — 使えなければ PTY に自動 fallback',
-  'settings.codexDelivery.optPty': 'PTY 注入 — 常にターミナルへ貼り付け',
-  'settings.codexDelivery.hint':
-    'codex への team_send を、codex 公式 app-server (JSON-RPC) 経由で送るか、従来どおりターミナルへ PTY 注入するか。バックエンドは履歴に残り入力競合も避けられますが、app-server が使えない場合は自動で PTY に fallback します。Windows は app-server 未対応のため常に PTY です。',
-  'settings.section.roles.label': 'ロール定義',
-  'settings.section.roles.title': 'ロール定義',
-  'settings.section.roles.desc': 'チームメンバーの役割テンプレ',
-  'settings.section.mcp.label': 'MCP',
-  'settings.section.mcp.title': 'MCP',
-  'settings.section.mcp.desc': 'vibe-team MCP の導入方法',
-  // Issue #825: 音声指揮モード (Voice Direction, Beta)
-  'settings.section.voice.label': '音声指揮 (Beta)',
-  'settings.section.voice.title': '音声指揮',
-  'settings.section.voice.desc': 'OpenAI Realtime API で AI と会話して Leader を指揮する',
-  'settings.voice.beta.warning':
-    'この機能はベータで、動作テストを行っていません。意図しない挙動・不安定な接続・誤認識が発生する可能性があります。フィードバックは GitHub Issue でお寄せください。',
-  'settings.voice.enabled.label': '音声指揮を有効化',
-  'settings.voice.apiKey.label': 'API キー',
-  'settings.voice.apiKey.placeholder': 'sk-...',
-  'settings.voice.apiKey.save': '保存',
-  'settings.voice.apiKey.clear': 'クリア',
-  'settings.voice.apiKey.clearConfirm': 'API キーを削除しますか?',
-  'settings.voice.apiKey.savedNotice':
-    'API キーは OS のキーリング (Windows: 資格情報マネージャー / macOS: キーチェーン / Linux: secret-service) に暗号化して安全に保存しています。一度保存すると再表示されません。再入力する場合は「クリア」してください。',
-  'settings.voice.model.label': 'モデル',
-  'settings.voice.voiceName.label': 'AI の声',
-  'settings.voice.language.label': '言語',
-  'settings.voice.inputDevice.label': '入力デバイス (マイク)',
-  'settings.voice.outputDevice.label': '出力デバイス (スピーカー)',
-  'settings.voice.shortcut.label': 'トグルショートカット',
-  'settings.voice.shortcut.reset': 'リセット',
-  'settings.voice.shortcut.capturing': '入力中… (キーを押してください)',
-  'settings.voice.confirmation.label': '送信時の確認',
-  'settings.voice.confirmation.always': '毎回確認する (推奨)',
-  'settings.voice.confirmation.bypass': '確認を省略する (バイパス)',
-  'settings.voice.confirmation.bypassWarning':
-    'バイパス時は AI からの音声確認も Renderer 側の最終確認もスキップされ、誤認識でも即座に Leader へ送信されます。',
-  'settings.voice.disclaimer.title': '音声指揮 (Beta)',
-  'settings.voice.disclaimer.body':
-    'この機能はベータで、開発者による動作テストを行っていません。意図しない挙動が発生する可能性があることをご了承ください。\n\n以下を理解した上でご利用ください:\n- OpenAI Realtime API を使用します。API 料金が発生します。\n- API キーは OS のキーリングに暗号化して保管されます。\n- マイクへのアクセス許可が必要です。\n- 認識精度や接続安定性は環境に依存します。\n- 不具合や改善要望は GitHub Issue でお寄せください。',
-  'settings.voice.disclaimer.ack': '理解しました',
-  'voice.button.idle': 'クリックで会話開始',
-  'voice.button.connecting': '接続中…',
-  'voice.button.listening': '会話中 — クリックで終了',
-  'voice.button.disabled.noKey': '設定で API キーを保存してください',
-  'voice.button.disabled.notEnabled': '設定で音声指揮を有効化してください',
-  'voice.confirm.title': '危険な操作の確認',
-  'voice.confirm.body': '次のメッセージを Leader に送信しますか?\n\n「{text}」',
-  'voice.confirm.send': '送信する',
-  'voice.confirm.cancel': 'キャンセル',
-  'voice.trail.sending': 'Leader へ送信中… (3 秒後に確定)',
-  'voice.trail.spawningTeam': 'チームを起動中… ({preset}, 3 秒後に確定)',
-  'voice.trail.cancel': 'キャンセル',
-  'voice.toast.apiKeySaved': 'API キーを保存しました',
-  'voice.toast.apiKeyCleared': 'API キーを削除しました',
-  'voice.toast.sent': 'Leader に送信しました',
-  'voice.toast.sendFailed': '送信に失敗しました ({code})',
-  'voice.error.micDenied': 'マイクへのアクセスが拒否されました',
-  'voice.error.openai401': 'OpenAI 認証エラー (API キーを確認してください)',
-  'voice.error.keyringUnavailable': 'OS のキーリングが利用できません',
-  'common.show': '表示',
-  'common.hide': '隠す',
-  'common.saving': '保存中…',
-  'common.systemDefault': 'システム既定',
-  'settings.section.logs.label': 'ログ',
-  'settings.section.logs.title': 'ログ',
-  'settings.section.logs.desc': 'アプリの実行ログを表示',
-  'settings.section.untitled': '（無名）',
-  'settings.section.customDesc': 'カスタムエージェント設定',
-  'settings.section.addCustom': '+ 追加',
-  'settings.section.group.agents': 'エージェント',
-  'settings.section.group.team': 'チーム',
-  'settings.section.group.other': 'その他',
-  // Issue #729: SettingsModal の inline isJa を i18n.ts に移管
-  'settings.dialog.label': '設定',
-  'settings.back': '戻る',
-  'settings.sections.ariaLabel': '設定セクション',
-  'settings.saveFailedSeeConsole': '設定の保存に失敗しました。詳細は開発者ツールのコンソールを確認してください。',
-  'settings.search.placeholder': '設定を検索…',
-  'settings.search.ariaLabel': '設定を検索',
-  'settings.search.clear': 'クリア',
-  'settings.search.noMatches': '一致する項目がありません',
-  'settings.fonts.uiFontTitle': 'UI フォント',
-  'settings.fonts.editorFontTitle': 'エディタフォント (Monaco)',
-  'settings.launch.title': '起動オプション',
-  'settings.launch.argsLabel': '引数（空白区切り、ダブルクォートで空白を含む値）',
-  'settings.launch.argsLabelSimple': '引数（空白区切り）',
-  'settings.launch.cwdLabel': '作業ディレクトリ（空なら現在のプロジェクトルート）',
-  'settings.launch.cwdUnset': '（未設定）',
-  'settings.launch.applyNote': '変更後は再起動でターミナルに反映されます。',
-  'settings.customAgents.newName': '新しいエージェント',
-  'settings.language': '言語',
-  'settings.language.desc':
-    'UI 表示言語を切り替え。Claude Code 自体の応答言語には影響しません。',
-  'settings.theme': 'テーマ',
-  'settings.uiFont': 'UI フォント',
-  'settings.uiFontFamily': 'フォントファミリ',
-  'settings.uiFontSize': 'サイズ (px)',
-  'settings.editorFont': 'エディタフォント (Monaco)',
-  'settings.editorFontFamily': 'フォントファミリ',
-  'settings.editorFontSize': 'サイズ (px)',
-  'settings.terminal': 'ターミナル',
-  'settings.terminalFontFamily': 'フォント',
-  'settings.terminalFontSize': 'フォントサイズ (px)',
-  'settings.terminalNote':
-    '既定は JetBrains Mono Nerd Font (本体同梱)。Powerline / Devicons / Material Icons の glyph を含み、Starship や oh-my-posh の icon が tofu になりません。★ は本体にバンドルされたフォントで、OS 未インストールでも常に同じルックで描画されます。',
-  'settings.terminalForceUtf8.label': 'Windows ターミナルで UTF-8 を強制 (chcp 65001)',
-  'settings.terminalForceUtf8.hint':
-    'cmd.exe / PowerShell 起動時に chcp 65001 を inject して console output を UTF-8 化します。漢字ファイル名や日本語出力が U+FFFD 化するのを防ぎます。OEM コードページを意図的に使いたい場合のみ OFF にしてください。Windows 以外の OS では何もしません。',
-  'settings.terminalForceUtf8.nonWindows': 'この設定は Windows でのみ有効です',
-  'settings.density': '情報密度',
-  // Issue #729: DensitySection 旧 hardcoded JP desc を i18n.ts に移管 (theme.desc / mascot.desc と同型)
-  'density.desc.compact': '14"以下の画面向け、余白小',
-  'density.desc.normal': '既定',
-  'density.desc.comfortable': '大画面向け、ゆったり',
-  'settings.reset': 'デフォルトに戻す',
-  'settings.cancel': 'キャンセル',
-  'settings.apply': '適用して保存',
-  'settings.custom': '（カスタム）',
-
-  // ---------- Theme labels (UserMenu / OnboardingWizard 共有) ----------
-  'theme.label.claude-dark': 'Claude Dark',
-  'theme.label.claude-light': 'Claude Light',
-  'theme.label.dark': 'ダーク',
-  'theme.label.light': 'ライト',
-  'theme.label.midnight': 'ミッドナイト',
-  'theme.label.glass': 'グラス',
-
-  // ---------- Theme descriptions (ThemeSection の theme card 用) ----------
-  // Issue #729: 旧 settings-options.ts の hardcoded JP `desc` を i18n.ts に移管。EN ユーザー向け表示を修正。
-  'theme.desc.claude-dark': 'Anthropic 公式カラー準拠。ウォームダークブラウン + コーラル #D97757（既定）',
-  'theme.desc.claude-light': 'claude.ai のクリーム背景と温かい差し色を再現',
-  'theme.desc.dark': 'VS Code 系のクラシックダーク',
-  'theme.desc.midnight': '深い青紫ベース、紫アクセント',
-  'theme.desc.glass': 'すりガラス風 — 半透明パネル + ブラー',
-  'theme.desc.light': '明るい背景、暗い文字',
-
-  // ---------- Language labels (UserMenu / LanguageSection 共有) ----------
-  'lang.label.ja': '日本語',
-  'lang.label.ja.sub': 'Japanese',
-  'lang.label.en': 'English',
-  'lang.label.en.sub': 'English',
-
-  // ---------- Settings: Logs (Issue #326) ----------
-  'settings.logs.title': 'ログ',
-  'settings.logs.desc':
-    'アプリの実行ログ (~/.vibe-editor/logs/vibe-editor.log) の末尾を表示します。バグ報告にはこのログを添付してください。',
-  'settings.logs.refresh': '再読み込み',
-  'settings.logs.openDir': 'ログフォルダを開く',
-  'settings.logs.levelFilter': 'レベル',
-  'settings.logs.level.all': 'すべて',
-  'settings.logs.loading': '読み込み中…',
-  'settings.logs.empty': 'ログはまだありません。',
-  'settings.logs.noMatch': '選択したレベルに該当するログがありません。',
-  'settings.logs.truncated': '末尾のみ表示中',
-
-  // ---------- Toast ----------
-  'toast.reviewRequested': '差分レビューを依頼: {path}',
-  'toast.sessionResumed': 'セッションに復帰: {title}',
-  'toast.pathCopied': 'パスをクリップボードにコピー',
-  'toast.copyFailed': 'クリップボードへのコピーに失敗しました',
-  'toast.revealFailed': 'ファイルマネージャでの表示に失敗しました',
-  // Issue #592: ファイル操作のフィードバック
-  'toast.fileCreated': '"{name}" を作成しました',
-  'toast.folderCreated': 'フォルダ "{name}" を作成しました',
-  'toast.fileRenamed': '"{from}" を "{to}" にリネームしました',
-  'toast.fileDeleted': '"{name}" を削除しました',
-  'toast.fileCopied': '"{name}" をコピーしました',
-  'toast.fileMoved': '"{name}" を移動しました',
-  'toast.fileOpFailed': 'ファイル操作に失敗しました: {error}',
-  'toast.fileOpClipboardEmpty': 'クリップボードに対象がありません',
-  'toast.terminalNotReady': 'ターミナルが起動していません',
-  'toast.settings.loadFailed':
-    '設定ファイルを読み込めなかったため、この起動中は設定の自動保存を停止しました: {error}',
-  'toast.settings.saveBlocked':
-    '設定ファイルを読み込めなかったため、設定の保存を停止しています。アプリを再起動してください。',
-  'toast.settings.saveFailed': '設定の保存に失敗しました: {error}',
-  'toast.settings.projectRootFailed': 'プロジェクトルートの反映に失敗しました: {error}',
-  // Issue #578: Canvas 非表示中に recruit が走った件数を可視化時に警告する
-  'toast.recruitWhileHidden':
-    'Canvas を非表示の間にメンバー採用が {count} 件走りました。失敗していたら再実行してください',
-  'toast.recruitRescued': '採用 (遅着救済): {ms}ms 遅れて受領されました',
-
-  // ---------- Terminal (pasteエラー等) ----------
-  'terminal.pasteImageFailed': '画像保存失敗',
-  'terminal.pasteException': 'ペースト例外',
-
-  // ---------- Terminal cwd warning (Issue #818) ----------
-  // Rust 側 `resolve_valid_cwd` が無効 cwd で fallback したとき、warning を
-  // 日本語ハードコードせず i18n key + params で renderer に渡す (#729 取り残し対応)。
-  // - `{requested}`: 指定された cwd (空文字なら下記 `*.unsetLabel` を埋める)
-  // - `{fallback}` : フォールバック先 (project root か process default)
-  'terminal.cwd.warningPrefix': '[警告]',
-  'terminal.cwd.unsetLabel': '(未設定)',
-  'terminal.cwd.invalidFallbackToHome':
-    '指定された作業ディレクトリが無効です: {requested} → {fallback} で起動します',
-  'terminal.cwd.invalidFallbackToProcessDefault':
-    '作業ディレクトリが無効です: {requested} → プロセス既定の {fallback} で起動します',
-
-  // ---------- Terminal context menu (Issue #356) ----------
-  'terminal.ctxMenu.paste': '貼り付け',
-  'terminal.ctxMenu.copySelection': '選択範囲をコピー',
-  'terminal.ctxMenu.clear': 'ターミナルをクリア',
-
-  // ---------- Command palette (Issue #39) ----------
-  'palette.ariaLabel': 'コマンドパレット',
-  'palette.placeholder': 'コマンドを検索…',
-  'palette.hint': '↑↓ で選択 · Enter で実行 · Esc で閉じる',
-  'palette.count': '{count} 件',
-  'palette.empty': '一致するコマンドがありません',
-
-  // ---------- Canvas QuickNav (Issue #58) ----------
-  'quicknav.placeholder': 'エージェント / カードへジャンプ…',
-  'quicknav.empty': '該当するカードがありません',
-  'quicknav.hintNavigate': '↑↓ 選択',
-  'quicknav.hintJump': 'Enter ジャンプ',
-  'quicknav.hintClose': 'Esc 閉じる',
-
-  // ---------- Command palette entries (Issue #57) ----------
-  'cmd.cat.project': 'プロジェクト',
-  'cmd.cat.workspace': 'ワークスペース',
-  'cmd.cat.view': 'ビュー',
-  'cmd.cat.tab': 'タブ',
-  'cmd.cat.git': 'Git',
-  'cmd.cat.sessions': 'セッション',
-  'cmd.cat.terminal': 'ターミナル',
-  'cmd.cat.settings': '設定',
-  'cmd.cat.theme': 'テーマ',
-  'cmd.project.new': '新規プロジェクト…',
-  'cmd.project.openFolder': 'フォルダを開く…',
-  'cmd.project.openFile': 'ファイルを開く…',
-  'cmd.workspace.addFolder': 'フォルダをワークスペースに追加…',
-  'cmd.project.recent': '最近: {name}',
-  'cmd.view.sidebarChanges': 'サイドバー: 変更',
-  'cmd.view.sidebarSessions': 'サイドバー: 履歴',
-  'cmd.view.nextTab': '次のタブへ',
-  'cmd.view.prevTab': '前のタブへ',
-  'cmd.tab.close': 'アクティブなタブを閉じる',
-  'cmd.tab.reopen': '最近閉じたタブを復元',
-  'cmd.tab.togglePin': 'アクティブなタブをピン留め/解除',
-  'cmd.git.refresh': '変更ファイル一覧を更新',
-  'cmd.sessions.refresh': 'セッション履歴を更新',
-  'cmd.terminal.addClaude': 'Claude Code タブを追加',
-  'cmd.terminal.addCodex': 'Codex タブを追加',
-  'cmd.terminal.closeTab': 'アクティブなターミナルタブを閉じる',
-  'cmd.terminal.restart': 'ターミナルを再起動',
-
-  // ---------- Terminal pane (exit handling) ----------
-  'terminal.exited': '終了',
-  'terminal.exitedTitle': 'プロセスが終了しています',
-  'terminal.exitedBanner': 'プロセスが終了しました ({status})',
-  'customAgent.warn.args': 'カスタムエージェントの引数に解析上の警告があります（未閉じクォート / 特殊ダッシュ）。設定を確認してください',
-  'customAgent.warn.modelOverride': '明示モデル指定（{model}）で起動します。プランで利用できないモデルの場合、claude が API エラーを繰り返すことがあります',
-  'terminal.apiErrorHint': 'API エラーが繰り返し発生しています。起動したモデルがプランで利用できない可能性があります。/model で標準モデルに変更してください',
-  'terminal.status.starting': '{command} を起動中…',
-  'terminal.status.running': '実行中: {command}',
-  'terminal.status.exited': '終了 (exitCode={exitCode})',
-  'terminal.status.spawnFailed': '起動失敗: {error}',
-  'terminal.status.reconnect': '再接続: {command}',
-  'terminal.status.reconnectRestored': '再接続 (出力復元): {command}',
-  'terminal.status.exception': '例外: {error}',
-  'terminal.limitReached': 'ターミナル上限（{max}）に達しました',
-  'terminal.limitWarning': 'ターミナル数が {threshold} に達しました（上限 {max}）',
-  'terminal.restart': '再起動',
-  'terminal.closeTab': '閉じる',
-  'layout.sidebarResizeTitle': 'ドラッグでサイドバー幅を調整 / ダブルクリックでリセット',
-  'layout.idePanelResizeTitle': 'ドラッグで IDE モードパネルの幅を調整',
-  'cmd.settings.open': '設定を開く',
-  'cmd.settings.cycleDensity': '情報密度を切り替え',
-  'cmd.settings.cycleDensitySub': '現在: {density}',
-  'cmd.theme.title': 'テーマ: {name}',
-  'cmd.theme.current': '✓ 現在のテーマ',
-  'cmd.cat.app': 'アプリ',
-  'cmd.app.restart': 'vibe-editor (アプリ) を再起動',
-
-  // ---------- Settings 補助 (Issue #76) ----------
-  'settings.command': 'コマンド',
-  'settings.argsUnterminatedQuote': 'ダブルクォート (") が閉じていません。引数が誤って解釈される可能性があります。',
-  'settings.argsUnicodeDash':
-    'Unicode ダッシュ (–, — など) が含まれています。実行時に ASCII の "--" に自動変換します。コピペや IME の自動変換が原因の可能性があります。',
-
-  // ---------- Custom agents ----------
-  'settings.customAgents.add': '+ カスタムエージェントを追加',
-  'settings.agentWizard.title': 'エージェントを追加',
-  'settings.agentWizard.typeApi': 'API モデル',
-  'settings.agentWizard.typeApiDesc': 'OpenAI / Anthropic / Gemini などのクラウドモデル',
-  'settings.agentWizard.typeCli': 'CLI コマンド',
-  'settings.agentWizard.typeCliDesc': 'ターミナルで動く CLI エージェント (claude/codex 互換)',
-  'settings.agentWizard.apiKeyOptional': 'API キー (任意・後で設定可)',
-  'settings.agentWizard.skills': 'Skill',
-  'settings.agentWizard.skillsHint':
-    '起動時に効かせる skill を選択します (任意)。未 import の場合は設定の Skill セクションから取り込めます。',
-  'settings.agentWizard.reviewSummary': '以下の内容で作成します',
-  'settings.agentWizard.next': '次へ',
-  'settings.agentWizard.back': '戻る',
-  'settings.agentWizard.create': '作成',
-  'settings.agentWizard.cancel': 'キャンセル',
-  'settings.customAgents.name': '表示名',
-  'settings.customAgents.remove': '削除',
-  'settings.customAgents.untitled': '（無名）',
-  // Issue #729: CustomAgentEditor の isJa 三項を i18n.ts に集約
-  'settings.customAgents.confirmDelete': 'カスタムエージェント "{name}" を削除しますか？',
-  'settings.customAgents.namePlaceholder': '例: Aider',
-  'settings.customAgents.argsLabel': '引数（空白区切り、ダブルクォートで空白を含む値）',
-  'settings.customAgents.cwdLabel': '作業ディレクトリ（空なら現在のプロジェクトルート）',
-  'settings.customAgents.cwdUnset': '（未設定）',
-  'settings.customAgents.accentColor': 'アクセントカラー（任意）',
-  'settings.customAgents.engine': 'engine（互換系統）',
-  'settings.customAgents.engineClaude': 'Claude 互換',
-  'settings.customAgents.engineCodex': 'Codex 互換',
-  'settings.customAgents.icon': 'アイコン（lucide 名）',
-  'settings.customAgents.tags': 'タグ',
-  'settings.customAgents.tagsPlaceholder': 'カンマ区切り',
-  'settings.customAgents.runtime': '実行方式',
-  'settings.customAgents.provider': 'Provider',
-  'settings.customAgents.baseUrl': 'Base URL',
-  'settings.customAgents.model': 'Model',
-  'settings.customAgents.apiKey': 'API key',
-  'settings.customAgents.apiKeySaved': '保存済み（値は表示されません）',
-  'settings.customAgents.apiKeyClearConfirm': '保存済み API key を削除しますか？',
-  'settings.customAgents.apiKeySaveError': 'API キーの保存に失敗しました: {detail}',
-  'settings.customAgents.toolMode': 'Tool mode',
-  'settings.customAgents.toolAuto': 'Auto',
-  'settings.customAgents.toolReadOnly': 'Read-only chat',
-  'settings.customAgents.systemPrompt': 'System prompt override',
-  'settings.customAgents.apiNote': 'TeamHub tool は provider/model が対応する場合のみ有効です。',
-  'settings.customAgents.readOnlyNote':
-    'この provider/model は tool calling を read-only chat に degrade します。',
-  'settings.customAgents.applyNote': '変更後、Canvas で該当エージェントのカードを作り直すと反映されます。',
-  'settings.customAgents.skills': 'Skill (SKILL.md)',
-  'settings.customAgents.skillsEmpty':
-    'import 済みの skill がありません。下の「Claude / Codex から import」で追加してください。',
-  'settings.customAgents.skillsAutoTeam': 'TeamHub 参加時は vibe-team skill が自動で追加されます。',
-  'settings.customAgents.skillSearch': 'skill を検索',
-  'settings.customAgents.applySkills': 'プロジェクトに適用',
-  'settings.customAgents.applySkillsBusy': '適用中…',
-  'settings.customAgents.applySkillsEmpty': '適用する skill が選択されていません。',
-  'settings.customAgents.applySkillsDone':
-    '{count}/{total} 件の skill を .claude/skills に適用しました。',
-  'settings.customAgents.applySkillsError': 'skill の適用に失敗しました: {detail}',
-  'settings.customAgents.cliSkillsNote':
-    'CLI エージェントは .claude/skills を自動探索します。チェックして「適用」でプロジェクトに配置してください。',
-  'settings.customAgents.skillImport.title': 'Claude / Codex から skill を import',
-  'settings.customAgents.skillImport.note':
-    '~/.claude/skills と ~/.agents/skills (Codex) を走査し、選んだ skill を vibe-editor 専用フォルダにコピーします。',
-  'settings.customAgents.skillImport.empty':
-    'import 元 (~/.claude/skills・~/.agents/skills) に skill が見つかりません。',
-  'settings.customAgents.skillImport.import': 'Import',
-  'settings.customAgents.skillImport.remove': '削除',
-
-  // ---------- MCP tab ----------
-  'settings.mcp.autoTitle': '自動セットアップ',
-  'settings.mcp.autoLabel': 'Team 起動時に vibe-team MCP を自動で登録する',
-  'settings.mcp.autoHint':
-    '~/.claude.json や ~/.codex/config.toml を書き換えます。書き込みに失敗する場合は OFF にして、下の手順で自分で入れてください。',
-  'settings.mcp.aiTitle': 'AI エージェントに入れさせる',
-  'settings.mcp.aiDesc':
-    '以下のプロンプトを Claude Code / Codex に貼り付けて実行させると、vibe-team MCP がセットアップされます。',
-  'settings.mcp.manualTitle': '手動で入れる',
-  'settings.mcp.manualDesc': '好みのエディタで設定ファイルを開いて、以下の断片をマージしてください。',
-  'settings.mcp.manualStep1': '~/.claude.json を開く (無ければ新規作成)。',
-  'settings.mcp.manualStep2': '最上位の "mcpServers" オブジェクトに "vibe-team" エントリを追加。',
-  'settings.mcp.manualStep3': 'Codex を使う場合は ~/.codex/config.toml に同等の [mcp_servers.vibe-team] を追加。',
-  'settings.mcp.copy': 'コピー',
-  'settings.mcp.copied': 'コピーしました',
-  // Issue #729: McpSection の isJa 三項を i18n.ts に移管
-  'settings.mcp.claudeSampleNote': '~/.claude.json のサンプル (既存の mcpServers と統合してください):',
-  'settings.mcp.codexSampleNote': '~/.codex/config.toml のサンプル:',
-  'settings.mcp.connInfoLabel': '接続情報 (現在値):',
-
-  // ---------- Updater (Issue #59) ----------
-  'updater.confirm': 'vibe-editor v{version} が利用可能です。今すぐ更新しますか?',
-  'updater.upToDate': '最新版を使用しています',
-  'updater.checkFailed': '更新の確認に失敗しました: {error}',
-  'updater.dialogFailed': '更新ダイアログの表示に失敗しました: {error}',
-  'updater.downloading': '更新をダウンロード中…',
-  'updater.downloadProgress': 'ダウンロード中… {pct}%',
-  'updater.installing': 'インストール中… 完了後に再起動します',
-  'updater.downloadFailed': 'ダウンロードに失敗しました: {error}',
-  'updater.relaunchFailed': '再起動に失敗しました ({error})。手動で再起動してください',
-  'updater.runningTasksWarning': '実行中のエージェントが {count} 個あります。更新で中断されます',
-  'updater.checkNow': '更新を確認',
-  'updater.button.label': '更新 v{version}',
-  'updater.button.title': '新しいバージョン v{version} が利用可能です。クリックでインストール',
-  // Issue #609: minisign 署名検証失敗の警告 (24h に 1 度だけ表示)
-  'updater.signatureFailed':
-    '更新ファイルの署名検証に失敗しました。改竄や中継経路の異常の可能性があります。次回更新までしばらくお待ちください。',
-
-  // ---------- Toast tone ラベル (Issue #80) ----------
-  'toast.tone.info': '情報',
-  'toast.tone.success': '完了',
-  'toast.tone.warning': '注意',
-  'toast.tone.error': 'エラー',
-
-  // ---------- Terminal タブ復元 (Issue #857) ----------
-  'terminalTabs.restore.transcriptMissing':
-    '過去の会話履歴が見つからず {count} 件のタブを新規会話で再起動しました',
-  'terminalTabs.saveFailed':
-    'ターミナルタブの保存に失敗しました: {error}',
-
-  // ---------- Status ----------
-  'status.noProject': 'プロジェクトが選択されていません',
-
-  // ---------- Image preview ----------
-  'imagePreview.devUnavailable': 'dev:vite モードでは画像プレビューを利用できません。',
-  'imagePreview.loadError': '画像を表示できません: {path}',
-
-  // ---------- Team history ----------
-  'teamHistory.resume.emptyMembers': 'チームメンバー情報が空のため復元できません',
-  'teamHistory.resume.otherProject':
-    'このチームは別プロジェクト({project})の履歴です',
-  'teamHistory.resume.terminalLimit':
-    'ターミナル上限({max})を超えるため復元できません',
-
-  // ---------- Onboarding ----------
-  'onboarding.back': '戻る',
-  'onboarding.next': '次へ',
-  'onboarding.skip': 'あとでにする',
-  'onboarding.replay': 'セットアップをもう一度',
-  'onboarding.ariaLabel': 'vibe-editor セットアップ',
-  'onboarding.welcome.eyebrow': 'vibe-editor',
-  'onboarding.welcome.title': '静かな集中の、新しい入口。',
-  'onboarding.welcome.subtitle':
-    'Claude Code と Codex のための、穏やかな IDE。数ステップだけ、ご一緒させてください。',
-  'onboarding.welcome.cta': 'はじめる',
-  'onboarding.appearance.eyebrow': 'Appearance',
-  'onboarding.appearance.title': '見た目を選ぶ',
-  'onboarding.appearance.subtitle': '言語とテーマは、あとで設定からいつでも変えられます。',
-  'onboarding.appearance.language': '言語',
-  'onboarding.appearance.theme': 'テーマ',
-  'onboarding.workspace.eyebrow': 'Workspace',
-  'onboarding.workspace.title': '最初のフォルダを開く',
-  'onboarding.workspace.subtitle':
-    'プロジェクトの場所を選ぶと、次回以降も自動で開きます。あとから追加してもかまいません。',
-  'onboarding.workspace.choose': 'フォルダを選ぶ',
-  'onboarding.workspace.change': '別のフォルダを選ぶ',
-  'onboarding.workspace.clear': '選択したフォルダをクリア',
-  'onboarding.done.eyebrow': 'Ready',
-  'onboarding.done.title': '準備ができました',
-  'onboarding.done.subtitle': '落ち着いた画面で、今日の一行を書きはじめましょう。',
-  'onboarding.done.summaryLanguage': '言語',
-  'onboarding.done.summaryTheme': 'テーマ',
-  'onboarding.done.summaryFolder': 'フォルダ',
-  'onboarding.done.summaryFolderNone': 'あとで開く',
-  'onboarding.done.cta': 'エディタを開く'
-};
-
-const en: Dict = {
-  // ---------- Common ----------
-  'common.close': 'Close',
-  'common.cancel': 'Cancel',
-
-  // ---------- Toolbar ----------
-  'toolbar.restart.title': 'Restart app',
-  'toolbar.palette.title': 'Command palette (Ctrl+Shift+P)',
-  'toolbar.settings.title': 'Settings (Ctrl+,)',
-
-  // ---------- Window controls (Issue #260 PR-2: custom titlebar) ----------
-  'windowControls.minimize': 'Minimize',
-  'windowControls.maximize': 'Maximize',
-  'windowControls.restore': 'Restore',
-  'windowControls.close': 'Close',
-  'windowControls.group': 'Window controls',
-
-  // ---------- Topbar (redesign shell) ----------
-  'topbar.mode.canvas': 'Canvas',
-
-  // ---------- Status bar ----------
-  'status.branch': 'branch',
-  'status.changes': 'changes',
-  'status.lang': 'lang',
-  'status.theme': 'theme',
-  'status.mascot.idle': 'Idle',
-  'status.mascot.sleep': 'Sleeping…',
-  'status.mascot.working': 'Agent working',
-  'status.mascot.thinking': 'Waiting for response',
-  'status.mascot.done': 'Done!',
-  'status.mascot.error': 'Needs attention',
-  'status.mascot.excited': 'Yeah!',
-
-  // ---------- Mascot section (SettingsModal "Character" section) ----------
-  // Issue #729: MascotSection isJa ternaries / settings-options.ts hardcode -> centralised in i18n.ts
-  'settings.mascot.title': 'Character',
-  'settings.mascot.pickTitle': 'Pick a mascot image',
-  'settings.mascot.imageFilterName': 'Images',
-  'settings.mascot.choose': 'Choose image…',
-  'settings.mascot.clear': 'Clear',
-  'settings.mascot.hint':
-    'PNG / GIF (animated) / APNG / WebP / SVG. A small square (32–128 px) renders best.',
-  'mascot.desc.vibe': 'Default tiny companion',
-  'mascot.desc.spark': 'Brighter and lighter',
-  'mascot.desc.mono': 'A terminal-friendly angular look',
-  'mascot.desc.coder': 'A tiny companion typing at a computer',
-  'mascot.desc.custom': 'Use your own image (PNG/GIF/SVG/WebP) as the companion',
-
-  // ---------- Canvas HUD ----------
-  'canvas.apiAgent.teamRole': 'Team role',
-  'canvas.apiAgent.teamRolePlaceholder': 'e.g. reviewer',
-  'canvas.apiChat.placeholder': 'Type an instruction (/ for commands, @ to pick an agent)',
-  'canvas.apiChat.typing': '{name} is typing…',
-  'canvas.apiChat.ready': 'Ready. Enter your instruction.',
-  'canvas.apiChat.loadingPrompt': 'Loading system prompt…',
-  'canvas.apiChat.configure': 'Configure this API agent in Settings.',
-  'canvas.apiChat.mention': 'Mention an agent (@)',
-  'canvas.apiChat.attach': 'Attach file (coming soon)',
-  'canvas.apiChat.send': 'Send',
-  'canvas.apiChat.stop': 'Stop',
-  'canvas.apiChat.cmd.planDesc': 'Make a plan',
-  'canvas.apiChat.cmd.statusDesc': 'Check status',
-  'canvas.apiChat.cmd.contextDesc': 'Show context',
-  'canvas.apiChat.cmd.clearDesc': 'Clear history',
-  'canvas.hud.stage': 'Stage',
-  'canvas.hud.list': 'List',
-  'canvas.hud.focus': 'Focus',
-  'canvas.hud.fit': 'Fit',
-  'canvas.hud.zoomIn': 'Zoom in',
-  'canvas.hud.zoomOut': 'Zoom out',
-  'canvas.hud.arrange.open': 'Arrange',
-  // Issue #368: hover tooltips (Label — purpose)
-  'canvas.hud.stage.tooltip': 'Stage — Switch to a radial layout of agents',
-  'canvas.hud.list.tooltip': 'List — Show agents stacked vertically',
-  'canvas.hud.focus.tooltip': 'Focus — Highlight only the selected agent',
-  'canvas.hud.fit.tooltip': 'Fit — Auto-zoom so every card on the canvas fits the viewport',
-  'canvas.hud.zoomIn.tooltip': 'Zoom in — Enlarge the canvas',
-  'canvas.hud.zoomOut.tooltip': 'Zoom out — Shrink the canvas',
-  'canvas.hud.arrange.open.tooltip': 'Arrange — Tidy cards, unify size, and adjust spacing',
-  'canvas.hud.arrange.tidy': 'Tidy up',
-  'canvas.hud.arrange.unifySize': 'Unify size',
-  'canvas.hud.arrange.gap.label': 'Gap',
-  'canvas.hud.arrange.gap.tight': 'Tight',
-  'canvas.hud.arrange.gap.normal': 'Normal',
-  'canvas.hud.arrange.gap.wide': 'Wide',
-
-  // ---------- AppMenu ----------
-  'appMenu.title': 'Project menu',
-  'appMenu.new': 'New project…',
-  'appMenu.newHint': 'Create or select empty folder',
-  'appMenu.openFolder': 'Open folder…',
-  'appMenu.openFolderHint': 'Existing project',
-  'appMenu.openFile': 'Open file…',
-  'appMenu.newDialogTitle': 'New project',
-  'appMenu.openFolderDialogTitle': 'Open folder',
-  'appMenu.openFileDialogTitle': 'Open file',
-  'project.newDialogTitle': 'New project: choose or create an empty folder',
-  'project.openExistingDialogTitle': 'Open existing project',
-  'project.loading': 'Loading project…',
-  'project.loadError': 'Load error: {error}',
-  'project.initError': 'Initialization error: {error}',
-  'project.newFolderNotEmpty': 'Folder is not empty. Opening it as an existing project.',
-  'project.created': 'Created new project',
-  'project.fileParentLoaded': 'Loaded the parent folder of {file} as the project',
-  'project.recentCleared': 'Cleared recent project history',
-  'appMenu.addWorkspaceDialogTitle': 'Add to workspace',
-  'appMenu.openFileHint': 'Single file',
-  'appMenu.addToWorkspace': 'Add folder to workspace…',
-  'appMenu.addToWorkspaceHint': 'Show another root in the sidebar',
-  'appMenu.recent': 'Recent projects',
-  'appMenu.recentCount': '{count} recent',
-  'appMenu.workspace': 'Workspace',
-  'appMenu.clear': 'Clear',
-  'appMenu.empty': 'No history',
-  'menubar.file': 'File',
-  'menubar.view': 'View',
-  'menubar.help': 'Help',
-  'menubar.toggleSidebar': 'Toggle sidebar',
-  'menubar.toggleCanvas': 'Toggle IDE / Canvas',
-  'menubar.openPalette': 'Command palette',
-  'menubar.openSettings': 'Settings…',
-  'menubar.openGithub': 'Open on GitHub',
-  'menubar.restart': 'Restart',
-  // ---------- UserMenu (sidebar footer) ----------
-  'userMenu.settings': 'Settings',
-  'userMenu.language': 'Language',
-  'userMenu.theme': 'Theme',
-  'userMenu.releases': 'View releases on GitHub',
-  // ---------- Workspace (Issue #4) ----------
-  'workspace.roots': 'Workspace',
-  'workspace.add': 'Add folder',
-  'workspace.remove': 'Remove from workspace',
-  'workspace.removePrimaryConfirm': 'Remove {name} from the current workspace?',
-  'workspace.removed': 'Removed {name} from the workspace',
-  'workspace.added': 'Added {name} to the workspace',
-  'workspace.alreadyAdded': '{name} is already in the workspace',
-
-  // ---------- Sidebar ----------
-  'sidebar.files': 'Files',
-  'sidebar.changes': 'Changes',
-  'sidebar.history': 'History',
-  'sidebar.loading': 'Loading…',
-  'sidebar.notGitRepo': 'Not a git repository',
-  'sidebar.noChanges': 'No changes',
-  'sidebar.noSessions': 'No session history for this project yet',
-  'sidebar.filesChanged': '{count} changed',
-  'sidebar.sessionCount': '{count} sessions',
-  'sidebar.refresh': 'Refresh',
-  'sidebar.teams': 'Teams',
-  'sidebar.singleSessions': 'Single sessions',
-  'sidebar.notes': 'Notes',
-  'rail.primaryNav': 'Primary navigation',
-
-  // ---------- Notes (Issue #17) ----------
-  'notes.title': 'Notes',
-  'notes.placeholder': 'Jot down anything you want to hand off between terminals…\nSaved automatically.',
-  'notes.copy': 'Copy to clipboard',
-  'notes.clear': 'Clear notes',
-  'notes.copied': 'Copied notes',
-  'notes.copyFailed': 'Failed to copy',
-  'notes.confirmClear': 'Clear notes?',
-  'notes.autoSaved': 'Saved automatically',
-  'notes.chars': 'chars',
-
-  // ---------- Team history ----------
-  'teamHistory.resume': 'Resume team "{name}"',
-  'teamHistory.resumed': 'Resumed team "{name}"',
-  'teamHistory.alreadyOpen': 'Team "{name}" is already open on the Canvas',
-  'teamHistory.delete': 'Remove from history',
-
-  // ---------- File tree / Editor ----------
-  'filetree.refresh': 'Reload',
-  'filetree.treeLabel': 'File tree',
-  'diff.loading': 'Loading diff…',
-  'diff.selectFile': 'Select a file to view its diff',
-  'diff.error': 'Error: {error}',
-  'diff.binary': 'Binary files cannot be shown as diffs: {path}',
-  'diff.new': '(new)',
-  'diff.deleted': '(deleted)',
-  'diff.toggleMode': 'Toggle diff display mode',
-  'diff.toggleInline': 'Switch to inline',
-  'diff.toggleSideBySide': 'Switch to side by side',
-  'editor.loading': 'Loading file…',
-  'editor.save': 'Save (Ctrl+S)',
-  'editor.save.ariaLabel': 'Save',
-  'editor.viewPreview': 'Show preview',
-  'editor.viewSource': 'Show source',
-  'editor.binaryNotice': 'Binary file cannot be edited: {path}',
-  'editor.nonUtf8Warning':
-    'Opened with lossy encoding ({path}) — saving would lose the original encoding so editing is disabled.',
-  'editor.nonUtf8SaveBlocked': 'Save is disabled (non-UTF-8): {path}',
-  'editor.nonUtf8ReadOnly': 'read-only (non-UTF-8)',
-  'editor.externalChangeConfirm':
-    '{path} has been modified on disk since you opened it. Save anyway and overwrite external changes?',
-  'editor.saveAborted': 'Save aborted: {path}',
-  'editor.saved': 'Saved: {path}',
-  'editor.saveFailed': 'Save failed: {error}',
-  'editor.discardSingle': 'This file has unsaved changes. Close it anyway?\n\n{path}',
-  'editor.discardMultiple': 'There are unsaved changes. Switching now will discard {count} file(s). Continue?',
-  'editor.restartConfirm': 'There are unsaved changes. Restarting the app will discard them. Continue?',
-  // Issue #595: Confirmation shown when closing a Canvas EditorCard with unsaved edits via × / Clear.
-  'editor.confirmDiscardChanges':
-    'This card has unsaved changes that will be lost if you close it. Continue?\n\n{path}',
-  'editor.confirmDiscardChangesPlural':
-    '{count} cards have unsaved changes that will be lost if you close them. Continue?\n\n{paths}',
-
-  // ---------- Welcome ----------
-  'welcome.subtitle': 'vibe coding with Claude Code',
-  'welcome.hint1Key': 'Right',
-  'welcome.hint1Text': "talk to Claude Code in the terminal",
-  'welcome.hint2Key': 'Changes',
-  'welcome.hint2Text': "tab: review diffs of files Claude touched",
-  'welcome.hint3Key': 'History',
-  'welcome.hint3Text': 'tab: resume past sessions',
-  'welcome.hint4Text': 'for the command palette',
-
-  // ---------- Context menu ----------
-  'ctxMenu.openDiff': 'Open diff',
-  'ctxMenu.reviewDiff': 'Ask Claude Code to review this diff',
-  'ctxMenu.copyPath': 'Copy path',
-  // Issue #251: file tree right-click menu
-  'ctxMenu.copyAbsolutePath': 'Copy absolute path',
-  'ctxMenu.copyRelativePath': 'Copy relative path',
-  'ctxMenu.copyFileName': 'Copy file name',
-  'ctxMenu.revealInFolder': 'Reveal in File Explorer',
-  // Issue #592: VS Code-style file/folder operations
-  'ctxMenu.newFile': 'New File',
-  'ctxMenu.newFolder': 'New Folder',
-  'ctxMenu.rename': 'Rename',
-  'ctxMenu.delete': 'Delete',
-  'ctxMenu.cut': 'Cut',
-  'ctxMenu.copy': 'Copy',
-  'ctxMenu.paste': 'Paste',
-  'ctxMenu.duplicate': 'Duplicate',
-  'filetree.prompt.newFileName': 'New file name',
-  'filetree.prompt.newFolderName': 'New folder name',
-  'filetree.prompt.renameTo': 'New name',
-  'filetree.confirmDeleteFile': 'Move "{name}" to the trash?',
-  'filetree.confirmDeleteFolder': 'Move "{name}" and all of its contents to the trash?',
-  'filetree.confirmDeletePermanent': 'Permanently delete "{name}"? This action cannot be undone.',
-  'filetree.preloadRestartRequired': 'Restart the app to apply the preload update',
-  'canvasMenu.lockTeam': 'Move team together',
-  'canvasMenu.unlockTeam': 'Unlock team movement',
-  'canvasMenu.deleteCard': 'Delete card',
-  'canvasMenu.addClaudeHere': 'Add Claude here',
-  'canvasMenu.addCodexHere': 'Add Codex here',
-  'canvasMenu.addCustomAgentHere': 'Add {name} here',
-  'canvasMenu.addFileTreeHere': 'Add file tree here',
-  'canvasMenu.addChangesHere': 'Add Git changes here',
-  'canvasMenu.addEditorHere': 'Add empty editor here',
-  'canvasMenu.spawnDefaultTeam': 'Spawn default team',
-
-  // ---------- Claude Code panel ----------
-  'claudePanel.title': 'IDE Mode',
-  'claudePanel.notFound.title': 'Claude Code not found',
-  'claudePanel.notFound.body':
-    'The `claude` command was not found on your PATH. Install Claude Code, or specify the launch command in Settings.',
-  'claudePanel.notFound.step1Title': 'Install the CLI',
-  'claudePanel.notFound.step1Desc': 'Make sure the `claude` command is available from your terminal.',
-  'claudePanel.notFound.step2Title': 'Check settings',
-  'claudePanel.notFound.step2Desc': 'If using a custom command, review the launch command in Settings.',
-  'claudePanel.notFound.installLink': 'Install Claude Code',
-  'claudePanel.notFound.retry': 'Retry detection',
-  'claudePanel.notFound.settings': 'Open settings',
-  'claudePanel.checking': 'Checking…',
-  'claudePanel.newTab': 'New terminal tab',
-  'claudePanel.addClaude': 'Add Claude Code',
-  'claudePanel.addCodex': 'Add Codex',
-
-  // ---------- Team ----------
-  'team.closeTeamConfirm': 'This is the team leader. Close entire team?',
-  'team.closeTeam': 'Close Team',
-  'team.closeLeaderOnly': 'Close Leader Only',
-
-  // ---------- Canvas ----------
-  'canvas.spawnTeam': 'Spawn Team',
-  'canvas.spawnTeam.tooltip': 'Spawn Team — Launch leader and members in one click using the default preset',
-  'canvas.spawnTeamMore': 'More team options…',
-  'canvas.spawnTeamMore.tooltip': 'Pick a preset — Choose another built-in preset or a recently used team',
-  'canvas.preset': 'Preset',
-  'canvas.preset.leaderClaude': 'Leader only (Claude Code)',
-  'canvas.preset.leaderCodex': 'Leader only (Codex)',
-  'canvas.preset.builtinHeader': 'Built-in',
-  'canvas.preset.savedHeader': 'Saved',
-  'canvas.preset.leaderCustom': 'Leader only ({name})',
-  'canvas.recent': 'Recent',
-  'canvas.noRecentTeams': 'No recent teams. Start one from a preset.',
-  'canvas.clear': 'Clear',
-  'canvas.clear.tooltip': 'Clear — Remove every card from the canvas',
-  'canvas.clearConfirm': 'Clear every card on the canvas?',
-  // Issue #595: Shown when Clear is invoked while one or more EditorCards have unsaved edits.
-  'canvas.clearConfirmWithDirtyEditors':
-    'Clearing the canvas will discard {count} unsaved edit(s). Continue?\n\n{paths}',
-  'canvas.switchToIde': 'Switch to IDE mode',
-  'canvas.switchToIde.tooltip': 'IDE — Return to the editor + terminal IDE mode',
-  'canvas.modeToggle': 'Switch to Canvas mode',
-  'canvas.card.editor': 'Editor',
-  'canvas.list.title': 'Team',
-  'canvas.list.empty': 'No agents have been placed yet',
-
-  // ---------- Agent Card ----------
-  'agentCard.close': 'Close card',
-  'agentCard.confirmCloseTeam':
-    'Closing this card will also close all {count} members of team "{name}" (including the Leader). Continue?',
-  'handoff.create': 'Hand off',
-  'handoff.createTooltip':
-    'Save a handoff document and ask the leader to recruit a successor and switch over via MCP',
-  'handoff.created': 'Handoff saved ({file}); MCP instructions sent to the leader PTY',
-  'handoff.action.reveal': 'Reveal saved file',
-  'handoff.error.noProject':
-    'Project root is not set. Open a folder from the sidebar, then try again.',
-  'handoff.error.createFailed': 'Failed to create handoff: {detail}',
-  'handoff.error.notLeader': 'Handoff can only be initiated from a Leader card',
-  'handoff.error.injectFailed': 'Failed to inject the MCP instructions into the leader PTY: {detail}',
-  // Issue #511: PTY inject failure warning + manual retry
-  'injectFailure.title': 'Delivery failed ({code}): {message}',
-  'injectFailure.retry': 'Retry',
-  'injectFailure.retryBusy': 'Retrying…',
-  'injectFailure.retrySuccess': 'Message re-delivered successfully',
-  'injectFailure.retryFailed': 'Retry failed ({reason})',
-  'injectFailure.retryError': 'Error during retry: {detail}',
-  'injectFailure.dismiss': 'Dismiss',
-  // Issue #509: delivered-but-not-read message indicator
-  'inboxUnread.label': '{count} unread ({ageSec}s elapsed)',
-  'inboxUnread.tooltip':
-    'This agent has {count} delivered message(s) that have not been confirmed via team_read for {ageSec} seconds. Consider nudging if it exceeds 60s.',
-  'agentStatus.idle': 'Idle',
-  'agentStatus.thinking': 'Thinking',
-  'agentStatus.typing': 'Typing',
-
-  // Issue #521: Agent card 3-line summary
-  'agentCard.summary.region': 'Agent status summary',
-  'agentCard.summary.noTask': 'No task assigned',
-  'agentCard.summary.needsLeader': 'Awaiting leader input',
-  'agentCard.summary.ago.unobserved': 'No output observed yet',
-  'agentCard.summary.ago.now': 'Output just now',
-  'agentCard.summary.ago.sec': 'Last output {value}s ago',
-  'agentCard.summary.ago.min': 'Last output {value}m ago',
-  'agentCard.summary.ago.hour': 'Last output {value}h ago',
-  'agentCard.summary.ago.day': 'Last output {value}d ago',
-
-  // Issue #510: Agent card health badge (sourced from TeamHub diagnostics)
-  'agentCard.summary.health.state.alive': 'Alive',
-  'agentCard.summary.health.state.stale': 'Stale',
-  'agentCard.summary.health.state.dead': 'Unresponsive',
-  'agentCard.summary.health.state.unknown': 'Unknown',
-  'agentCard.summary.health.silent.sec': '{state} (silent for {value}s)',
-  'agentCard.summary.health.silent.min': '{state} (silent for {value}m)',
-  'agentCard.summary.health.tooltip': 'Health: {state} · last self-status: {status}',
-  'agentCard.summary.health.noStatus': 'no self-reported status',
-
-  // Issue #521: Canvas-wide summary HUD
-  'canvas.hud.summary.label': 'Canvas team summary',
-  'canvas.hud.summary.active': 'Active',
-  'canvas.hud.summary.active.tooltip': 'Active — agents with recent output',
-  'canvas.hud.summary.blocked': 'Awaiting leader',
-  'canvas.hud.summary.blocked.tooltip':
-    'Awaiting leader — agents waiting for leader input or handoff ack',
-  'canvas.hud.summary.stale': 'Stale',
-  'canvas.hud.summary.stale.tooltip': 'Stale — agents with no output for 5+ minutes',
-  'canvas.hud.summary.completed': 'Completed',
-  'canvas.hud.summary.completed.tooltip':
-    'Completed — agents with acked handoff or retired sessions',
-  'canvas.hud.summary.dead': 'Unresponsive',
-  'canvas.hud.summary.dead.tooltip':
-    'Unresponsive — agents with no PTY output for 15+ minutes (sourced from hub diagnostics)',
-
-  // Issue #522: Team Presets panel
-  'preset.title': 'Team Presets',
-  'preset.button.tooltip': 'Presets — save and reapply team formations',
-  'preset.saveCurrent': 'Save current team',
-  'preset.saveCurrent.tooltip': 'Save the agent cards currently on the canvas as a preset',
-  'preset.save': 'Save',
-  'preset.name': 'Name',
-  'preset.namePlaceholder': 'e.g. Plan + Build + Review team',
-  'preset.description': 'Description',
-  'preset.descriptionPlaceholder': 'Optional notes (what kind of work this team is suited for)',
-  'preset.apply': 'Apply',
-  'preset.apply.tooltip': 'Spawn this preset onto the canvas',
-  'preset.delete': 'Delete',
-  'preset.delete.tooltip': 'Delete this preset from disk',
-  'preset.empty': 'No saved presets yet',
-  'preset.loading': 'Loading…',
-  'preset.roleCount': '{count} roles',
-  'preset.saved': 'Preset "{name}" saved',
-  'preset.applied': 'Added {count} members from "{name}" to the canvas',
-  'preset.deleted': 'Preset "{name}" deleted',
-  'preset.error.empty': 'No agent cards on the canvas. Build a team first, then save it as a preset.',
-  'preset.error.noName': 'Please enter a preset name',
-  'preset.error.listFailed': 'Failed to load preset list',
-  'preset.error.saveFailed': 'Failed to save preset: {detail}',
-  'preset.error.deleteFailed': 'Failed to delete preset: {detail}',
-
-  // Issue #514: Team Dashboard
-  'dashboard.title': 'Team Dashboard',
-  'dashboard.button.tooltip':
-    'Team dashboard — overview of every member with state, task, and last activity',
-  'dashboard.count': '{count} members',
-  'dashboard.col.member': 'Member',
-  'dashboard.col.state': 'State',
-  'dashboard.col.task': 'Task',
-  'dashboard.col.lastSeen': 'Last seen',
-  'dashboard.state.active': 'Active',
-  'dashboard.state.blocked': 'Awaiting leader',
-  'dashboard.state.stale': 'Stale',
-  'dashboard.state.completed': 'Completed',
-  'dashboard.state.idle': 'Idle',
-  'dashboard.task.unassigned': 'No task assigned',
-  'dashboard.lastSeen.never': 'never',
-  'dashboard.empty.noTeam':
-    'No agent team on this canvas. Add at least one agent card to use the dashboard.',
-  'dashboard.empty.noMembers':
-    'This team has no members yet. Recruit members from the Leader using `team_recruit`.',
-  'dashboard.banner.humanGate': 'Human gate blocked: waiting for leader decision',
-  'dashboard.alert.leaderInput': 'Awaiting Leader input',
-  'dashboard.alert.staleOutput': 'No output for 5+ minutes',
-  // Issue #615: dual / multi preset support for team section heading
-  'dashboard.team.label': 'Team {index}',
-
-  // ---------- Sessions ----------
-  'sessions.resume': 'Resume session {id}',
-  'sessions.messages': '{count} msgs',
-  // Issue #837: "N+" rendering when messageCount reaches the scan limit.
-  'sessions.messagesCapped': '{count}+ msgs',
-  'sessions.loadMore': 'Load {remaining} more',
-
-  // ---------- Tab ----------
-  'tab.pinned': 'Pinned',
-  'tab.newOutput': 'New output',
-  'tab.pin': 'Pin tab',
-  'tab.unpin': 'Unpin',
-  'tab.close': 'Close tab',
-  'tab.closeWithShortcut': 'Close (Ctrl+W)',
-  'fonts.family': 'Font family',
-  'fonts.custom': '(custom)',
-  'fonts.size': 'Size (px)',
-  'fonts.customCss': 'Custom CSS font-family',
-
-  // ---------- Roles ----------
-
-  // ---------- Settings ----------
-  'settings.title': 'Settings',
-  // Issue #729: WelcomePane inline isJa moved into i18n.ts
-  'welcome.title': 'Build with calm momentum.',
-  'welcome.recentProjects': 'Recent projects',
-  'welcome.recentProjectsTitle': 'Jump back into your flow',
-  'welcome.workspaceLabel': 'Workspace',
-  'welcome.quickStart': 'Quick start',
-  'welcome.quickStartTitle': 'What you can do next',
-  // Issue #729: canvas-layout-helpers language-based hardcode moved into i18n.ts
-  'canvas.agentCount': '{count} agents',
-  'canvas.orgAgentCount': '{organizationCount} orgs / {agentCount} agents',
-  // Issue #729: RoleProfilesSection inline isJa moved into i18n.ts
-  'settings.roles.title': 'Role profiles',
-  'settings.roles.desc':
-    'Define vibe-team member roles. Leaders pick from these when calling team_recruit.',
-  'settings.roles.globalPreamble': 'Global preamble',
-  'settings.roles.globalPreambleHint': 'Prepended to all prompts',
-  'settings.roles.confirmDelete': 'Delete "{id}"?',
-  'settings.roles.addCustom': 'Add custom role',
-  'settings.roles.newCustomDesc': 'New custom role.',
-  'settings.roles.builtin': 'built-in',
-  'settings.roles.custom': 'custom',
-  'settings.roles.color': 'Color',
-  'settings.roles.glyph': 'Glyph',
-  'settings.roles.defaultEngine': 'Default engine',
-  'settings.roles.permissions': 'Permissions',
-  'settings.roles.promptEn': 'System prompt (EN)',
-  'settings.roles.promptJa': 'System prompt (JA)',
-  'settings.roles.promptPlaceholders':
-    'Available: {teamName} {selfLabel} {selfDescription} {roster} {tools} {globalPreamble}',
-  'settings.roles.deleteRole': 'Delete this role',
-  // Issue #729: settings-section-meta.tsx FIXED_LABELS_EN moved into i18n.ts
-  'settings.section.general.label': 'General',
-  'settings.section.general.title': 'General',
-  'settings.section.general.desc': 'Language and density',
-  'settings.section.appearance.label': 'Appearance',
-  'settings.section.appearance.title': 'Appearance',
-  'settings.section.appearance.desc': 'Theme, surfaces, and character',
-  'settings.section.fonts.label': 'Fonts',
-  'settings.section.fonts.title': 'Typography',
-  'settings.section.fonts.desc': 'UI / editor / terminal fonts',
-  'settings.section.claude.label': 'Claude Code',
-  'settings.section.claude.title': 'Claude Code',
-  'settings.section.claude.desc': 'Launch command and args',
-  'settings.section.codex.label': 'Codex',
-  'settings.section.codex.title': 'Codex',
-  'settings.section.codex.desc': 'Launch command and args',
-  // Issue #1068: codex team_send delivery method toggle
-  'settings.codexDelivery.title': 'team_send delivery',
-  'settings.codexDelivery.label': 'Delivery method',
-  'settings.codexDelivery.optBackend': 'Backend (app-server) — falls back to PTY if unavailable',
-  'settings.codexDelivery.optPty': 'PTY injection — always paste into the terminal',
-  'settings.codexDelivery.hint':
-    'How team_send reaches codex: via the official codex app-server (JSON-RPC) or the legacy PTY paste into the terminal. Backend keeps history and avoids input races, but automatically falls back to PTY when the app-server is unavailable. Windows always uses PTY (app-server is not supported).',
-  'settings.section.roles.label': 'Role profiles',
-  'settings.section.roles.title': 'Role profiles',
-  'settings.section.roles.desc': 'Team member role templates',
-  'settings.section.mcp.label': 'MCP',
-  'settings.section.mcp.title': 'MCP',
-  'settings.section.mcp.desc': 'How to install vibe-team MCP',
-  // Issue #825: Voice Direction Mode (Beta)
-  'settings.section.voice.label': 'Voice (Beta)',
-  'settings.section.voice.title': 'Voice Direction',
-  'settings.section.voice.desc':
-    'Direct your Leader by talking to an AI assistant via OpenAI Realtime API.',
-  'settings.voice.beta.warning':
-    'This feature is in beta and has not been tested. Unexpected behavior, unstable connections, or misrecognition may occur. Please share feedback on GitHub Issues.',
-  'settings.voice.enabled.label': 'Enable voice direction',
-  'settings.voice.apiKey.label': 'API key',
-  'settings.voice.apiKey.placeholder': 'sk-...',
-  'settings.voice.apiKey.save': 'Save',
-  'settings.voice.apiKey.clear': 'Clear',
-  'settings.voice.apiKey.clearConfirm': 'Delete the saved API key?',
-  'settings.voice.apiKey.savedNotice':
-    'Your API key is encrypted and securely stored in your OS keyring (Credential Manager on Windows, Keychain on macOS, secret-service on Linux). Once saved, it cannot be viewed again. Click "Clear" to re-enter.',
-  'settings.voice.model.label': 'Model',
-  'settings.voice.voiceName.label': 'AI voice',
-  'settings.voice.language.label': 'Language',
-  'settings.voice.inputDevice.label': 'Input device (microphone)',
-  'settings.voice.outputDevice.label': 'Output device (speaker)',
-  'settings.voice.shortcut.label': 'Toggle shortcut',
-  'settings.voice.shortcut.reset': 'Reset',
-  'settings.voice.shortcut.capturing': 'Capturing… (press a key combination)',
-  'settings.voice.confirmation.label': 'Send confirmation',
-  'settings.voice.confirmation.always': 'Always confirm (recommended)',
-  'settings.voice.confirmation.bypass': 'Bypass confirmation',
-  'settings.voice.confirmation.bypassWarning':
-    'When bypassed, both the AI verbal confirmation and the renderer-side final check are skipped. Misrecognized speech may be sent to the Leader immediately.',
-  'settings.voice.disclaimer.title': 'Voice Direction (Beta)',
-  'settings.voice.disclaimer.body':
-    'This feature is in beta and has not been tested by the developers. Unexpected behavior may occur.\n\nPlease read before using:\n- It uses the OpenAI Realtime API. API charges apply.\n- Your API key is stored encrypted in your OS keyring.\n- Microphone permission is required.\n- Recognition accuracy and connection stability depend on your environment.\n- Please report issues and feedback on GitHub Issues.',
-  'settings.voice.disclaimer.ack': 'I understand',
-  'voice.button.idle': 'Click to start',
-  'voice.button.connecting': 'Connecting…',
-  'voice.button.listening': 'Listening — click to stop',
-  'voice.button.disabled.noKey': 'Save an API key in Settings',
-  'voice.button.disabled.notEnabled': 'Enable voice direction in Settings',
-  'voice.confirm.title': 'Confirm sensitive action',
-  'voice.confirm.body': 'Send the following message to the Leader?\n\n"{text}"',
-  'voice.confirm.send': 'Send',
-  'voice.confirm.cancel': 'Cancel',
-  'voice.trail.sending': 'Sending to Leader… (3 s before commit)',
-  'voice.trail.spawningTeam': 'Spawning team… ({preset}, 3 s before commit)',
-  'voice.trail.cancel': 'Cancel',
-  'voice.toast.apiKeySaved': 'API key saved',
-  'voice.toast.apiKeyCleared': 'API key cleared',
-  'voice.toast.sent': 'Sent to Leader',
-  'voice.toast.sendFailed': 'Send failed ({code})',
-  'voice.error.micDenied': 'Microphone access was denied',
-  'voice.error.openai401': 'OpenAI authentication error (check your API key)',
-  'voice.error.keyringUnavailable': 'OS keyring is not available',
-  'common.show': 'Show',
-  'common.hide': 'Hide',
-  'common.saving': 'Saving…',
-  'common.systemDefault': 'System default',
-  'settings.section.logs.label': 'Logs',
-  'settings.section.logs.title': 'Logs',
-  'settings.section.logs.desc': 'View runtime logs from the app',
-  'settings.section.untitled': '(untitled)',
-  'settings.section.customDesc': 'Custom agent settings',
-  'settings.section.addCustom': '+ Add',
-  'settings.section.group.agents': 'Agents',
-  'settings.section.group.team': 'Team',
-  'settings.section.group.other': 'Other',
-  // Issue #729: SettingsModal inline isJa moved into i18n.ts
-  'settings.dialog.label': 'Settings',
-  'settings.back': 'Back',
-  'settings.sections.ariaLabel': 'Settings sections',
-  'settings.saveFailedSeeConsole': 'Failed to save settings. See the developer console for details.',
-  'settings.search.placeholder': 'Search settings…',
-  'settings.search.ariaLabel': 'Search settings',
-  'settings.search.clear': 'Clear',
-  'settings.search.noMatches': 'No matches',
-  'settings.fonts.uiFontTitle': 'UI Font',
-  'settings.fonts.editorFontTitle': 'Editor Font (Monaco)',
-  'settings.launch.title': 'Launch options',
-  'settings.launch.argsLabel': 'Arguments',
-  'settings.launch.argsLabelSimple': 'Arguments',
-  'settings.launch.cwdLabel': 'Working directory',
-  'settings.launch.cwdUnset': '(unset)',
-  'settings.launch.applyNote': 'Restart terminals to apply changes.',
-  'settings.customAgents.newName': 'New agent',
-  'settings.language': 'Language',
-  'settings.language.desc':
-    'Switch the UI language. Does not affect the language Claude Code responds in.',
-  'settings.theme': 'Theme',
-  'settings.uiFont': 'UI font',
-  'settings.uiFontFamily': 'Font family',
-  'settings.uiFontSize': 'Size (px)',
-  'settings.editorFont': 'Editor font (Monaco)',
-  'settings.editorFontFamily': 'Font family',
-  'settings.editorFontSize': 'Size (px)',
-  'settings.terminal': 'Terminal',
-  'settings.terminalFontFamily': 'Font',
-  'settings.terminalFontSize': 'Font size (px)',
-  'settings.terminalNote':
-    'Default is JetBrains Mono Nerd Font (bundled). Includes Powerline / Devicons / Material Icons glyphs so Starship and oh-my-posh icons no longer render as tofu. ★ marks bundled fonts that always render the same regardless of OS-installed fonts.',
-  'settings.terminalForceUtf8.label': 'Force UTF-8 in Windows terminals (chcp 65001)',
-  'settings.terminalForceUtf8.hint':
-    'Inject `chcp 65001` when launching cmd.exe / PowerShell so console output is UTF-8. Prevents Japanese / CJK filenames and output from rendering as U+FFFD. Turn this OFF only if you intentionally want to keep the OEM code page. No-op on non-Windows OSes.',
-  'settings.terminalForceUtf8.nonWindows': 'This setting only applies on Windows',
-  'settings.density': 'Density',
-  // Issue #729: DensitySection hardcoded JP desc moved to i18n.ts (mirrors theme.desc / mascot.desc)
-  'density.desc.compact': 'For 14" or smaller screens, tighter spacing',
-  'density.desc.normal': 'Default',
-  'density.desc.comfortable': 'For large screens, roomier spacing',
-  'settings.reset': 'Reset to defaults',
-  'settings.cancel': 'Cancel',
-  'settings.apply': 'Apply & save',
-  'settings.custom': '(custom)',
-
-  // ---------- Theme labels (UserMenu / OnboardingWizard) ----------
-  'theme.label.claude-dark': 'Claude Dark',
-  'theme.label.claude-light': 'Claude Light',
-  'theme.label.dark': 'Dark',
-  'theme.label.light': 'Light',
-  'theme.label.midnight': 'Midnight',
-  'theme.label.glass': 'Glass',
-
-  // ---------- Theme descriptions (ThemeSection theme cards) ----------
-  // Issue #729: previously hardcoded JP in settings-options.ts. Now centralised so EN users see English.
-  'theme.desc.claude-dark': "Anthropic's official palette. Warm dark brown + coral #D97757 (default)",
-  'theme.desc.claude-light': 'Recreates the claude.ai cream background with warm accent colors',
-  'theme.desc.dark': 'Classic VS Code-style dark',
-  'theme.desc.midnight': 'Deep blue-purple base with purple accents',
-  'theme.desc.glass': 'Frosted-glass look — translucent panels + blur',
-  'theme.desc.light': 'Bright background, dark text',
-
-  // ---------- Language labels (UserMenu / LanguageSection) ----------
-  'lang.label.ja': '日本語',
-  'lang.label.ja.sub': 'Japanese',
-  'lang.label.en': 'English',
-  'lang.label.en.sub': 'English',
-
-  // ---------- Settings: Logs (Issue #326) ----------
-  'settings.logs.title': 'Logs',
-  'settings.logs.desc':
-    'Tail of the app runtime log (~/.vibe-editor/logs/vibe-editor.log). Attach this when filing a bug report.',
-  'settings.logs.refresh': 'Refresh',
-  'settings.logs.openDir': 'Open log folder',
-  'settings.logs.levelFilter': 'Level',
-  'settings.logs.level.all': 'All',
-  'settings.logs.loading': 'Loading…',
-  'settings.logs.empty': 'No logs yet.',
-  'settings.logs.noMatch': 'No log lines match the selected level.',
-  'settings.logs.truncated': 'tail only',
-
-  // ---------- Toast ----------
-  'toast.reviewRequested': 'Review requested: {path}',
-  'toast.sessionResumed': 'Resumed session: {title}',
-  'toast.pathCopied': 'Path copied to clipboard',
-  'toast.copyFailed': 'Failed to copy to clipboard',
-  'toast.revealFailed': 'Failed to reveal in file manager',
-  // Issue #592: file operation feedback
-  'toast.fileCreated': 'Created "{name}"',
-  'toast.folderCreated': 'Created folder "{name}"',
-  'toast.fileRenamed': 'Renamed "{from}" to "{to}"',
-  'toast.fileDeleted': 'Deleted "{name}"',
-  'toast.fileCopied': 'Copied "{name}"',
-  'toast.fileMoved': 'Moved "{name}"',
-  'toast.fileOpFailed': 'File operation failed: {error}',
-  'toast.fileOpClipboardEmpty': 'Nothing to paste',
-  'toast.terminalNotReady': 'Terminal is not ready',
-  'toast.settings.loadFailed':
-    'Failed to load settings, so automatic settings saves are disabled for this launch: {error}',
-  'toast.settings.saveBlocked':
-    'Settings were not loaded, so saving settings is disabled. Please restart the app.',
-  'toast.settings.saveFailed': 'Failed to save settings: {error}',
-  'toast.settings.projectRootFailed': 'Failed to apply project root: {error}',
-  // Issue #578: Warn when recruits ran while canvas was hidden
-  'toast.recruitWhileHidden':
-    '{count} recruit(s) ran while Canvas was hidden. Re-run any that may have failed',
-  'toast.recruitRescued': 'Recruit rescued after timeout ({ms}ms late)',
-
-  // ---------- Status ----------
-  // ---------- Terminal (paste errors) ----------
-  'terminal.pasteImageFailed': 'Paste image failed',
-  'terminal.pasteException': 'Paste exception',
-
-  // ---------- Terminal cwd warning (Issue #818) ----------
-  // Rust side `resolve_valid_cwd` returns a structured warning (i18n key + params)
-  // when the requested cwd is invalid and falls back to project root / process cwd.
-  // Previously Rust hardcoded a Japanese string which leaked through to EN users
-  // (Issue #729 leftover).
-  // - `{requested}`: the originally requested cwd (empty → use `*.unsetLabel`)
-  // - `{fallback}` : where we actually started (project root or process default)
-  'terminal.cwd.warningPrefix': '[warning]',
-  'terminal.cwd.unsetLabel': '(unset)',
-  'terminal.cwd.invalidFallbackToHome':
-    'The requested working directory is invalid: {requested} → starting in {fallback} instead',
-  'terminal.cwd.invalidFallbackToProcessDefault':
-    'Working directory is invalid: {requested} → starting in the process default {fallback} instead',
-
-  // ---------- Terminal context menu (Issue #356) ----------
-  'terminal.ctxMenu.paste': 'Paste',
-  'terminal.ctxMenu.copySelection': 'Copy selection',
-  'terminal.ctxMenu.clear': 'Clear terminal',
-
-  // ---------- Command palette (Issue #39) ----------
-  'palette.ariaLabel': 'Command palette',
-  'palette.placeholder': 'Search commands…',
-  'palette.hint': '↑↓ to select · Enter to run · Esc to close',
-  'palette.count': '{count}',
-  'palette.empty': 'No matching commands',
-
-  // ---------- Canvas QuickNav (Issue #58) ----------
-  'quicknav.placeholder': 'Jump to agent / card …',
-  'quicknav.empty': 'No matching cards.',
-  'quicknav.hintNavigate': '↑↓ navigate',
-  'quicknav.hintJump': 'Enter jump',
-  'quicknav.hintClose': 'Esc close',
-
-  // ---------- Command palette entries (Issue #57) ----------
-  'cmd.cat.project': 'Project',
-  'cmd.cat.workspace': 'Workspace',
-  'cmd.cat.view': 'View',
-  'cmd.cat.tab': 'Tab',
-  'cmd.cat.git': 'Git',
-  'cmd.cat.sessions': 'Sessions',
-  'cmd.cat.terminal': 'Terminal',
-  'cmd.cat.settings': 'Settings',
-  'cmd.cat.theme': 'Theme',
-  'cmd.project.new': 'New project…',
-  'cmd.project.openFolder': 'Open folder…',
-  'cmd.project.openFile': 'Open file…',
-  'cmd.workspace.addFolder': 'Add folder to workspace…',
-  'cmd.project.recent': 'Recent: {name}',
-  'cmd.view.sidebarChanges': 'Sidebar: Changes',
-  'cmd.view.sidebarSessions': 'Sidebar: History',
-  'cmd.view.nextTab': 'Next tab',
-  'cmd.view.prevTab': 'Previous tab',
-  'cmd.tab.close': 'Close active tab',
-  'cmd.tab.reopen': 'Reopen last closed tab',
-  'cmd.tab.togglePin': 'Toggle pin on active tab',
-  'cmd.git.refresh': 'Refresh changed files',
-  'cmd.sessions.refresh': 'Refresh session history',
-  'cmd.terminal.addClaude': 'Add Claude Code tab',
-  'cmd.terminal.addCodex': 'Add Codex tab',
-  'cmd.terminal.closeTab': 'Close active terminal tab',
-  'cmd.terminal.restart': 'Restart terminal',
-
-  // ---------- Terminal pane (exit handling) ----------
-  'terminal.exited': 'exited',
-  'terminal.exitedTitle': 'Process has exited',
-  'terminal.exitedBanner': 'Process exited ({status})',
-  'customAgent.warn.args': 'Custom agent args have parse warnings (unterminated quote / special dash). Please check the settings.',
-  'customAgent.warn.modelOverride': 'Launching with explicit model ({model}); if it is unavailable on your plan, claude may loop on API errors.',
-  'terminal.apiErrorHint': 'Repeated API errors detected. The launched model may be unavailable on your plan — switch to a standard model with /model.',
-  'terminal.status.starting': 'Starting {command}…',
-  'terminal.status.running': 'Running: {command}',
-  'terminal.status.exited': 'Exited (exitCode={exitCode})',
-  'terminal.status.spawnFailed': 'Start failed: {error}',
-  'terminal.status.reconnect': 'Reconnected: {command}',
-  'terminal.status.reconnectRestored': 'Reconnected (restored output): {command}',
-  'terminal.status.exception': 'Exception: {error}',
-  'terminal.limitReached': 'Terminal limit reached ({max})',
-  'terminal.limitWarning': 'Terminal count reached {threshold} (limit {max})',
-  'terminal.restart': 'Restart',
-  'terminal.closeTab': 'Close',
-  'layout.sidebarResizeTitle': 'Drag to resize the sidebar / double-click to reset',
-  'layout.idePanelResizeTitle': 'Drag to resize the IDE mode panel',
-  'cmd.settings.open': 'Open settings',
-  'cmd.settings.cycleDensity': 'Cycle density',
-  'cmd.settings.cycleDensitySub': 'Current: {density}',
-  'cmd.theme.title': 'Theme: {name}',
-  'cmd.theme.current': '✓ current theme',
-  'cmd.cat.app': 'App',
-  'cmd.app.restart': 'Restart vibe-editor',
-
-  // ---------- Settings helpers (Issue #76) ----------
-  'settings.command': 'Command',
-  'settings.argsUnterminatedQuote':
-    'Unterminated double quote (") — arguments may be parsed incorrectly.',
-  'settings.argsUnicodeDash':
-    'Contains Unicode dashes (–, — etc.) — they will be normalized to ASCII "--" at runtime. Likely caused by paste or IME autocorrect.',
-
-  // ---------- Custom agents ----------
-  'settings.customAgents.add': '+ Add custom agent',
-  'settings.agentWizard.title': 'Add an agent',
-  'settings.agentWizard.typeApi': 'API model',
-  'settings.agentWizard.typeApiDesc': 'Cloud models (OpenAI / Anthropic / Gemini …)',
-  'settings.agentWizard.typeCli': 'CLI command',
-  'settings.agentWizard.typeCliDesc': 'A CLI agent that runs in a terminal (claude/codex compatible)',
-  'settings.agentWizard.apiKeyOptional': 'API key (optional — can set later)',
-  'settings.agentWizard.skills': 'Skills',
-  'settings.agentWizard.skillsHint':
-    'Select skills to apply at launch (optional). Import more from the Skill section in settings.',
-  'settings.agentWizard.reviewSummary': 'Will create with:',
-  'settings.agentWizard.next': 'Next',
-  'settings.agentWizard.back': 'Back',
-  'settings.agentWizard.create': 'Create',
-  'settings.agentWizard.cancel': 'Cancel',
-  'settings.customAgents.name': 'Display name',
-  'settings.customAgents.remove': 'Remove',
-  'settings.customAgents.untitled': '(untitled)',
-  // Issue #729: CustomAgentEditor isJa ternaries consolidated into i18n.ts
-  'settings.customAgents.confirmDelete': 'Delete custom agent "{name}"?',
-  'settings.customAgents.namePlaceholder': 'e.g. Aider',
-  'settings.customAgents.argsLabel': 'Arguments (space-separated; use quotes for spaces)',
-  'settings.customAgents.cwdLabel': 'Working directory (blank = current project root)',
-  'settings.customAgents.cwdUnset': '(unset)',
-  'settings.customAgents.accentColor': 'Accent color (optional)',
-  'settings.customAgents.engine': 'Engine (compatibility)',
-  'settings.customAgents.engineClaude': 'Claude-compatible',
-  'settings.customAgents.engineCodex': 'Codex-compatible',
-  'settings.customAgents.icon': 'Icon (lucide name)',
-  'settings.customAgents.tags': 'Tags',
-  'settings.customAgents.tagsPlaceholder': 'Comma-separated',
-  'settings.customAgents.runtime': 'Runtime',
-  'settings.customAgents.provider': 'Provider',
-  'settings.customAgents.baseUrl': 'Base URL',
-  'settings.customAgents.model': 'Model',
-  'settings.customAgents.apiKey': 'API key',
-  'settings.customAgents.apiKeySaved': 'Saved (value is hidden)',
-  'settings.customAgents.apiKeyClearConfirm': 'Delete the saved API key?',
-  'settings.customAgents.apiKeySaveError': 'Failed to save the API key: {detail}',
-  'settings.customAgents.toolMode': 'Tool mode',
-  'settings.customAgents.toolAuto': 'Auto',
-  'settings.customAgents.toolReadOnly': 'Read-only chat',
-  'settings.customAgents.systemPrompt': 'System prompt override',
-  'settings.customAgents.apiNote': 'TeamHub tools are enabled only when the provider/model supports them.',
-  'settings.customAgents.readOnlyNote':
-    'This provider/model degrades to read-only chat because tool calling is unavailable.',
-  'settings.customAgents.applyNote': 'Recreate the agent card in Canvas to apply changes.',
-  'settings.customAgents.skills': 'Skills (SKILL.md)',
-  'settings.customAgents.skillsEmpty':
-    'No imported skills yet. Add some via “Import from Claude / Codex” below.',
-  'settings.customAgents.skillsAutoTeam':
-    'The vibe-team skill is added automatically when joining TeamHub.',
-  'settings.customAgents.skillSearch': 'Search skills',
-  'settings.customAgents.applySkills': 'Apply to project',
-  'settings.customAgents.applySkillsBusy': 'Applying…',
-  'settings.customAgents.applySkillsEmpty': 'No skills selected to apply.',
-  'settings.customAgents.applySkillsDone': 'Applied {count}/{total} skills to .claude/skills.',
-  'settings.customAgents.applySkillsError': 'Failed to apply skills: {detail}',
-  'settings.customAgents.cliSkillsNote':
-    'CLI agents auto-discover .claude/skills. Check skills and Apply to place them in the project.',
-  'settings.customAgents.skillImport.title': 'Import skills from Claude / Codex',
-  'settings.customAgents.skillImport.note':
-    'Scans ~/.claude/skills and ~/.agents/skills (Codex), and copies the selected skill into the vibe-editor skills folder.',
-  'settings.customAgents.skillImport.empty':
-    'No skills found in the import sources (~/.claude/skills, ~/.agents/skills).',
-  'settings.customAgents.skillImport.import': 'Import',
-  'settings.customAgents.skillImport.remove': 'Remove',
-
-  // ---------- MCP tab ----------
-  'settings.mcp.autoTitle': 'Auto setup',
-  'settings.mcp.autoLabel': 'Automatically register vibe-team MCP when a team starts',
-  'settings.mcp.autoHint':
-    'Rewrites ~/.claude.json and ~/.codex/config.toml. If that is unreliable, turn it off and install the server manually below.',
-  'settings.mcp.aiTitle': 'Have your AI agent install it',
-  'settings.mcp.aiDesc':
-    'Paste the following prompt into Claude Code or Codex and let it install the vibe-team MCP for you.',
-  'settings.mcp.manualTitle': 'Install manually',
-  'settings.mcp.manualDesc':
-    'Open the config files in your editor and merge the snippets below.',
-  'settings.mcp.manualStep1': 'Open ~/.claude.json (create it if missing).',
-  'settings.mcp.manualStep2': 'Add a "vibe-team" entry under the top-level "mcpServers" object.',
-  'settings.mcp.manualStep3':
-    'For Codex, add the equivalent [mcp_servers.vibe-team] section to ~/.codex/config.toml.',
-  'settings.mcp.copy': 'Copy',
-  'settings.mcp.copied': 'Copied',
-  // Issue #729: McpSection inline isJa moved into i18n.ts
-  'settings.mcp.claudeSampleNote': 'Sample for ~/.claude.json (merge with existing mcpServers):',
-  'settings.mcp.codexSampleNote': 'Sample for ~/.codex/config.toml:',
-  'settings.mcp.connInfoLabel': 'Connection info:',
-
-  // ---------- Updater (Issue #59) ----------
-  'updater.confirm': 'vibe-editor v{version} is available. Install it now?',
-  'updater.upToDate': 'You are on the latest version',
-  'updater.checkFailed': 'Failed to check for updates: {error}',
-  'updater.dialogFailed': 'Failed to show update dialog: {error}',
-  'updater.downloading': 'Downloading update…',
-  'updater.downloadProgress': 'Downloading… {pct}%',
-  'updater.installing': 'Installing… The app will restart when finished',
-  'updater.downloadFailed': 'Download failed: {error}',
-  'updater.relaunchFailed': 'Relaunch failed ({error}). Please restart manually',
-  'updater.runningTasksWarning': '{count} agent(s) are still running and will be interrupted',
-  'updater.checkNow': 'Check for updates',
-  'updater.button.label': 'Update v{version}',
-  'updater.button.title': 'A new version v{version} is available. Click to install',
-  // Issue #609: minisign signature failure warning (shown at most once per 24h)
-  'updater.signatureFailed':
-    'Update signature verification failed. The download may have been tampered with or routed through a faulty mirror. Please wait for the next update.',
-
-  // ---------- Toast tone labels (Issue #80) ----------
-  'toast.tone.info': 'Info',
-  'toast.tone.success': 'Success',
-  'toast.tone.warning': 'Warning',
-  'toast.tone.error': 'Error',
-
-  // ---------- Terminal tab restore (Issue #857) ----------
-  'terminalTabs.restore.transcriptMissing':
-    "Couldn't find past transcripts; restarted {count} tab(s) as new conversations.",
-  'terminalTabs.saveFailed': 'Stopped saving terminal tabs: {error}',
-
-  'status.noProject': 'No project selected',
-
-  // ---------- Image preview ----------
-  'imagePreview.devUnavailable': 'Image preview is unavailable in dev:vite mode.',
-  'imagePreview.loadError': 'Unable to display image: {path}',
-
-  // ---------- Team history ----------
-  'teamHistory.resume.emptyMembers': 'Cannot resume because team member information is empty',
-  'teamHistory.resume.otherProject':
-    'This team history belongs to another project ({project})',
-  'teamHistory.resume.terminalLimit':
-    'Cannot resume because it would exceed the terminal limit ({max})',
-
-  // ---------- Onboarding ----------
-  'onboarding.back': 'Back',
-  'onboarding.next': 'Next',
-  'onboarding.skip': 'Skip for now',
-  'onboarding.replay': 'Run setup again',
-  'onboarding.ariaLabel': 'vibe-editor setup',
-  'onboarding.welcome.eyebrow': 'vibe-editor',
-  'onboarding.welcome.title': 'A calmer entry to deep work.',
-  'onboarding.welcome.subtitle':
-    'A quiet IDE tailored for Claude Code and Codex. Just a couple of steps to get going.',
-  'onboarding.welcome.cta': 'Get started',
-  'onboarding.appearance.eyebrow': 'Appearance',
-  'onboarding.appearance.title': 'Choose your look',
-  'onboarding.appearance.subtitle':
-    'Language and theme can be changed anytime from settings.',
-  'onboarding.appearance.language': 'Language',
-  'onboarding.appearance.theme': 'Theme',
-  'onboarding.workspace.eyebrow': 'Workspace',
-  'onboarding.workspace.title': 'Open your first folder',
-  'onboarding.workspace.subtitle':
-    'Pick a project folder and we will reopen it next time. You can always add more later.',
-  'onboarding.workspace.choose': 'Choose folder',
-  'onboarding.workspace.change': 'Choose a different folder',
-  'onboarding.workspace.clear': 'Clear selected folder',
-  'onboarding.done.eyebrow': 'Ready',
-  'onboarding.done.title': 'You are all set',
-  'onboarding.done.subtitle': 'A calm workspace for today’s first line.',
-  'onboarding.done.summaryLanguage': 'Language',
-  'onboarding.done.summaryTheme': 'Theme',
-  'onboarding.done.summaryFolder': 'Folder',
-  'onboarding.done.summaryFolderNone': 'Open later',
-  'onboarding.done.cta': 'Open editor'
-};
-
-const translations: Record<Language, Dict> = { ja, en };
-
-/**
- * React フック: 現在の言語設定に基づいた翻訳関数を返す。
- *
- * ```
- * const t = useT();
- * t('sidebar.changes');                    // "変更" or "Changes"
- * t('sidebar.filesChanged', { count: 3 }); // "3 変更" or "3 changed"
- * ```
- */
-export function useT(): (key: string, params?: Record<string, string | number>) => string {
-  const { settings } = useSettings();
-  const lang = settings.language ?? 'ja';
-  return (key: string, params?: Record<string, string | number>): string => {
-    const text = translations[lang]?.[key] ?? translations.ja[key] ?? key;
-    if (!params) return text;
-    return interpolate(text, params);
-  };
-}
-
-/**
- * Issue #176: String.prototype.replace の第 2 引数は `$&` `$1` `$$` 等を
- * 特殊置換シーケンスとして解釈する。Windows パスや正規表現サンプル等を
- * params に渡すと結果が壊れていた。`replace(re, fn)` の関数フォームなら
- * 戻り値は literal として扱われるので安全。
- */
-function interpolate(text: string, params: Record<string, string | number>): string {
-  return Object.entries(params).reduce(
-    (acc, [k, v]) =>
-      acc.replace(new RegExp(`\\{${k}\\}`, 'g'), () => String(v)),
-    text
-  );
-}
-
-/**
- * React コンテキスト外 (updater-check / timer callback など) から呼べる翻訳関数。
- * 言語を明示的に受け取るので、呼び出し元が settings.language を取って渡す必要がある。
- */
-export function translate(
-  lang: Language,
-  key: string,
-  params?: Record<string, string | number>
-): string {
-  const text = translations[lang]?.[key] ?? translations.ja[key] ?? key;
-  if (!params) return text;
-  // Issue #176: replace の関数フォームを使って `$` 特殊シーケンスを literal 化
-  return Object.entries(params).reduce(
-    (acc, [k, v]) =>
-      acc.replace(new RegExp(`\\{${k}\\}`, 'g'), () => String(v)),
-    text
-  );
-}
