+++ src/renderer/src/lib/i18n/types.ts
@@ -1,0 +1,2 @@
+/** フラットキー i18n 辞書の共通型 (Issue #1032 分割) */
+export type Dict = Record<string, string>;
